refactor(bus): split flow consumer by concern and name the lane binding
Break the 1150-line flow_consumer.go into provisioning, flow-control
protocol, retry scheduling, delivery, and lane-mode files; carry the
stream/subject/consumer triple as one laneBinding value through the
provisioning path. Fold the duplicated New Relic attribute and trace-header
blocks behind shared helpers, name normalizedPlacement's nil test, give
baseOptions a connectionIdentity, and move stream reconciliation onto a
streamReconciler. Test files split along the same seams; assertion
inventory unchanged (196 tests before and after).
