+++ pkg/bus/flow_consumer.go
@@ -5,48 +5,25 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"strconv"
 	"strings"
-	"sync"
-	"sync/atomic"
 	"time"
 
 	"ItsBagelBot/pkg/env"
 
-	"github.com/bytedance/sonic"
 	"github.com/nats-io/nats.go"
 	jsapi "github.com/nats-io/nats.go/jetstream"
 	"github.com/nats-io/nuid"
-	"go.uber.org/zap"
 )
 
 const (
 	flowControlHeartbeat = time.Second
 	flowMaxAckPending    = 20_000
-	flowRetryTimeout     = 5 * time.Second
 
 	// flowProvisionTimeout bounds an ordinary create/update. The replacement path
 	// gets its own, much longer budget: a delete that succeeds followed by a
 	// create that times out leaves the stream with no consumer at all.
 	flowProvisionTimeout = 5 * time.Second
 	laneReplaceTimeout   = 20 * time.Second
-
-	// The server pins this ack policy's heartbeat to exactly one second, so three
-	// missed beats is the earliest safe conclusion that the consumer behind this
-	// delivery subject is gone. A push consumer learns about deletion only from
-	// heartbeat loss: no 404 and no 409 is ever published to a push delivery
-	// subject in 2.14.3.
-	flowHeartbeatGrace = 3 * time.Second
-	flowWatchdogTick   = time.Second
-
-	// flowWedgeStreak is how many consecutive stalled heartbeats, with no delivery
-	// between any two of them, prove the flow-control conversation has gone
-	// one-way. The heartbeat watchdog structurally cannot see this failure: the
-	// server heartbeats a STALLED consumer once a second indefinitely, so
-	// lastControl stays fresh for as long as the lane is frozen. Ten seconds of
-	// stall with nothing delivered is past any window the server would reopen on
-	// its own.
-	flowWedgeStreak = 10
 
 	// flowInactiveThreshold garbage-collects this pod's own consumer once the pod
 	// is gone. nats-server 2.14.3 arms a delete timer whenever a push consumer's
@@ -56,195 +33,22 @@ const (
 	// deliberately far longer than a reconnect: a pod that reattaches inside the
 	// window keeps its consumer and its ack floor.
 	flowInactiveThreshold = 5 * time.Minute
-
-	// flowSessionResetGap is the backwards jump in consumer sequence that can only
-	// be a new server-side session. A delete + recreate restarts the consumer
-	// sequence at 1 while the stream sequence keeps climbing, so anything further
-	// back than a whole ack window is a reset rather than reordering.
-	flowSessionResetGap = flowMaxAckPending
-
-	// flowControlPendingBytes is the server's flow-control window for this ack
-	// policy: nats-server 2.14.3 holds JsFlowControlMaxPending (32 MiB) of
-	// delivered payload outstanding before it shuts the window until a response
-	// arrives.
-	flowControlPendingBytes = 32 << 20
-
-	// flowWireBytesFloor is the smallest lane delivery the queue is sized for. A
-	// live R3 run measured ~865 B on the wire per ingress event; sizing at 800
-	// deliberately rounds the derived depth UP, because the failure this guards
-	// against is a queue too shallow for the window, never one too deep.
-	flowWireBytesFloor = 800
-
-	// flowQueueDepth is how many deliveries the callback may hold. It is derived
-	// from the BYTE window and not from flowMaxAckPending, because the message
-	// brake is not what binds here: under AckFlowControl the server keeps no
-	// per-message pending set to stop at, it stops on the byte window, and 32 MiB
-	// of ~865 B deliveries ramps to roughly 38k messages in flight — nearly twice
-	// the 20k the message figure suggests. A queue cut to flowMaxAckPending
-	// therefore overflows during perfectly healthy delivery: one repro dropped
-	// 26k deliveries in three seconds that way, silently, because a drop here
-	// costs an event rather than the binding. The depth has to sit above the
-	// ramped window, so it is derived at the wire floor with the margin that
-	// implies.
-	flowQueueDepth = flowControlPendingBytes / flowWireBytesFloor
 )
 
-// Push status protocol (nats-server 2.14 consumer.go, nats.go jetstream/push.go).
-// A status delivery carries no payload; the fleet answers two of them:
+// laneBinding is the stream/subject/consumer triple every provisioning step
+// carries as a unit. The three are same-typed and were passed positionally,
+// where a transposed pair is a consumer filtered on the wrong subject or bound
+// to the wrong stream — both of which provision cleanly and then deliver
+// nothing.
 //
-//	100 "FlowControl Request"  arrives inline in the delivery stream, reply set
-//	                           to the server's current flow-control id.
-//	100 "Idle Heartbeat"       carries Nats-Consumer-Stalled only while a
-//	                           flow-control response is still outstanding.
-//
-// Under AckFlowControl the response headers ARE the acknowledgement: the server
-// reads Nats-Last-Consumer/Nats-Last-Stream off the response and advances the
-// replicated ack floor to that stream sequence in one operation.
-const (
-	statusHeader          = "Status"
-	descriptionHeader     = "Description"
-	lastConsumerHeader    = "Nats-Last-Consumer"
-	lastStreamHeader      = "Nats-Last-Stream"
-	consumerStalledHeader = "Nats-Consumer-Stalled"
-
-	controlStatus = "100"
-)
-
-// RetryCountHeader bounds the flow path's redelivery to a single hop. It is an
-// ordinary application header, which is exactly why it can carry the budget: the
-// message scheduler strips every Nats-* header from what it emits and carries
-// application headers through verbatim, so this count survives the hop while the
-// schedule's own control headers do not.
-const RetryCountHeader = "Bagelbot-Retry"
-
-const (
-	// retryLanePrefix is the fleet's retry namespace, captured by
-	// TWITCH_INGRESS_RETRY. Both the schedule row and the message it emits live
-	// under it, because nats-server requires a schedule's Nats-Schedule-Target to
-	// collide with a subject of the very stream that stores the schedule
-	// (stream.go, JSMessageSchedulesTargetInvalid): a schedule can never deliver
-	// into a different stream, so the retry hop cannot land back on the ingress
-	// lane itself.
-	retryLanePrefix = "twitch.ingress.retry."
-
-	// retryScheduleTTL expires the emitted retry, not the schedule row. A retry
-	// nobody consumed within this window is a stale chat event and must not be
-	// acted on late; the one-shot row purges itself as it fires.
-	//
-	// Both this and the delay below are derived from TWITCH_INGRESS.MaxAge, which
-	// is 10s and IS the lane's staleness policy — an event the broker would have
-	// evicted for being too old to answer does not become answerable by having
-	// failed once. The original pair (3s delay, 30s TTL) predates that window and
-	// let a retry land ~33s after its failure, on top of the event's own age. Kept
-	// well inside it instead: one hop, delivered inside the window, expired at
-	// half of it. TWITCH_INGRESS_RETRY.MaxAge is 2m and is unrelated — it bounds
-	// the schedule ROW, which has to outlive its own delay to fire at all.
-	retryScheduleTTL = "5s"
-
-	defaultFlowRetryDelay = time.Second
-)
-
-// isHotIngressLane limits receipt-level acknowledgement flow control to the
-// perishable high-rate event lanes. Stream/status control messages retain the
-// ordinary explicit-ACK contract because losing one while an application
-// process exits has a much larger blast radius than replaying a chat event.
-// Work-queue streams are excluded by the same rule: their retention deletes on
-// ack and therefore requires per-message explicit acknowledgement.
-//
-// Two streams qualify, not one, because the lanes are partitioned across two
-// RAFT leaders: premium stays on TWITCH_INGRESS and standard lives on
-// TWITCH_INGRESS_STANDARD. The subject test still decides which lane it is —
-// the stream test only decides that the subject is an ingress lane at all — so
-// a lane keeps its acknowledgement contract across the partition. Omitting the
-// second stream here is silent: the standard lane would simply fall back to
-// explicit acks and the receipt-level path would be premium-only.
-func isHotIngressLane(stream, subject string) bool {
-	if stream != TwitchIngressStream.Name && stream != TwitchIngressStandardStream.Name {
-		return false
-	}
-	return strings.HasSuffix(subject, ".premium") || strings.HasSuffix(subject, ".standard")
-}
-
-// FlowConsumeEnabled reports whether the hot ingress lanes bind receipt-level
-// flow-controlled consumers. NATS_CONSUME_FLOW=on selects them; anything else,
-// including unset, keeps the explicit-ACK subscriber.
-//
-// It ships off, and the default is the whole contract: this flag is set in no
-// manifest under deploy/, so every consumer takes it the moment its image rolls.
-// Turning it on changes three things at once for the hot lanes — the server stops
-// tracking per-message pending state and advances one replicated ack floor per
-// window, a handler error can no longer NAK and is instead scheduled onto
-// TWITCH_INGRESS_RETRY, and that stream has to exist and be drained. Enabling it
-// therefore means provisioning the retry lane in the owning service AND
-// subscribing it in the same change. A default of "on" would do the first two
-// and not the third: failures would be scheduled onto a lane nobody reads and
-// silently expire.
-func FlowConsumeEnabled() bool {
-	return env.Get("NATS_CONSUME_FLOW", "off") == "on"
-}
-
-// laneConsumeMode is the acknowledgement contract the hot ingress lanes bind
-// under. The three are genuinely different shapes, not settings of one shape:
-//
-//	flow      per-pod AckFlowControl push consumer. Every pod receives the WHOLE
-//	          lane, so a pod added by the autoscaler multiplies delivery and
-//	          handler work rather than dividing it.
-//	pull      one fleet-wide durable, fetched by every pod. The server hands each
-//	          message to exactly one pod, so a pod added divides the lane.
-//	explicit  the ordinary durable queue-group consumer with per-message acks.
-type laneConsumeMode string
-
-const (
-	laneModeFlow     laneConsumeMode = "flow"
-	laneModePull     laneConsumeMode = "pull"
-	laneModeExplicit laneConsumeMode = "explicit"
-)
-
-// consumeMode resolves the configured contract. The default is deliberately
-// unchanged: flow is what production runs, and this knob exists to A/B pull
-// against it rather than to switch the fleet over.
-//
-// NATS_CONSUME_FLOW=off predates NATS_CONSUME_MODE and still wins outright,
-// because it is set in deployed manifests as the kill switch back to explicit
-// acks — an operator reaching for it during an incident must not have to know
-// that a second variable exists. An unrecognised mode falls back to pull for
-// the same reason: a typo must not silently change the lane's shape.
-//
-// Pull is the default receipt-level shape. The flow consumer fans out per
-// consumer name, so every pod added by the autoscaler receives a full copy of
-// the lane and multiplies the broker's delivery egress; the shared-durable pull
-// consumer divides the lane instead (live-measured 2026-08-15: two pull
-// instances per stream drained a 180k/s two-stream load the fan-out shape
-// could not, at 1x delivery egress). Ordering stays intact where it matters:
-// multi-line responses ship as one TypeBatch event, so cross-pod interleave
-// only touches independent events. Flow remains selectable per deployment for
-// any lane later shown to need a per-pod cursor.
-func consumeMode() laneConsumeMode {
-	if !FlowConsumeEnabled() {
-		return laneModeExplicit
-	}
-	switch laneConsumeMode(env.Get("NATS_CONSUME_MODE", string(laneModePull))) {
-	case laneModeFlow:
-		return laneModeFlow
-	case laneModeExplicit:
-		return laneModeExplicit
-	default:
-		return laneModePull
-	}
-}
-
-// RetryLaneSubject is the retry lane a failed hot-lane event is scheduled onto.
-// It is the schedule's target subject, and the subject a consumer binds to pick
-// the retry back up.
-func RetryLaneSubject(lane string) string {
-	return retryLanePrefix + subjectLeaf(lane)
-}
-
-func subjectLeaf(subject string) string {
-	if cut := strings.LastIndex(subject, "."); cut >= 0 {
-		return subject[cut+1:]
-	}
-	return subject
+// The consumer name is the one this pod owns on that stream, so a config built
+// from a binding always carries it as its Name: ensureFlowConsumer looks the
+// live consumer up by the binding and updates the config it was handed, and
+// those two must name the same durable.
+type laneBinding struct {
+	stream   string
+	subject  string
+	consumer string
 }
 
 // flowConsumerConfig keeps the hot consumer R3 while removing per-message
@@ -274,18 +78,18 @@ func subjectLeaf(subject string) string {
 // creation-time position on update and resumes at the predecessor's ack floor on
 // replacement, falling back to DeliverNew rather than to an unknown inherited
 // policy when that floor is unavailable.
-func flowConsumerConfig(subject, name string) jsapi.ConsumerConfig {
+func flowConsumerConfig(lane laneBinding) jsapi.ConsumerConfig {
 	return jsapi.ConsumerConfig{
-		Name:           name,
-		Durable:        name,
+		Name:           lane.consumer,
+		Durable:        lane.consumer,
 		Description:    "ItsBagelBot R3 flow-controlled ingress lane consumer",
 		DeliverPolicy:  jsapi.DeliverNewPolicy,
 		AckPolicy:      jsapi.AckFlowControlPolicy,
 		MaxDeliver:     -1,
-		FilterSubject:  subject,
+		FilterSubject:  lane.subject,
 		ReplayPolicy:   jsapi.ReplayInstantPolicy,
 		MaxAckPending:  flowMaxAckPending,
-		DeliverSubject: "_INBOX.BAGEL." + subjectToken(name),
+		DeliverSubject: "_INBOX.BAGEL." + subjectToken(lane.consumer),
 		FlowControl:    true,
 		// The server rejects any other heartbeat for this ack policy: the same
 		// interval bounds the stalled-source timeout it shares with sourcing.
@@ -353,8 +157,8 @@ func consumerToken(name string) string {
 // Recovery resumes just past this process's receipt cursor rather than at the
 // server's last-sent position, so nothing that was pushed but never arrived is
 // skipped. An empty cursor keeps the creation default rather than guessing.
-func recoveryFlowConsumerConfig(subject, name string, position flowPosition) jsapi.ConsumerConfig {
-	desired := flowConsumerConfig(subject, name)
+func recoveryFlowConsumerConfig(lane laneBinding, position flowPosition) jsapi.ConsumerConfig {
+	desired := flowConsumerConfig(lane)
 	if position.stream == 0 {
 		return desired
 	}
@@ -407,7 +211,7 @@ func requiresConsumerReplacement(err error) bool {
 
 // ensureFlowConsumer provisions this pod's flow consumer and returns the
 // delivery subject to bind, which is the server's value whenever one exists.
-func ensureFlowConsumer(nc *nats.Conn, stream string, desired jsapi.ConsumerConfig) (string, error) {
+func ensureFlowConsumer(nc *nats.Conn, lane laneBinding, desired jsapi.ConsumerConfig) (string, error) {
 	js, err := jsapi.NewWithDomain(nc, JSDomain())
 	if err != nil {
 		return "", fmt.Errorf("bus: modern jetstream context: %w", err)
@@ -415,12 +219,12 @@ func ensureFlowConsumer(nc *nats.Conn, stream string, desired jsapi.ConsumerConf
 	ctx, cancel := context.WithTimeout(context.Background(), flowProvisionTimeout)
 	defer cancel()
 
-	info, err := flowConsumerInfo(ctx, js, stream, desired.Name)
+	info, err := flowConsumerInfo(ctx, js, lane)
 	if err != nil {
 		return "", err
 	}
 	if info == nil {
-		_, err = js.CreateOrUpdatePushConsumer(ctx, stream, desired)
+		_, err = js.CreateOrUpdatePushConsumer(ctx, lane.stream, desired)
 		return desired.DeliverSubject, err
 	}
 
@@ -430,7 +234,7 @@ func ensureFlowConsumer(nc *nats.Conn, stream string, desired jsapi.ConsumerConf
 	desired.DeliverSubject = info.Config.DeliverSubject
 	desired.DeliverPolicy = info.Config.DeliverPolicy
 	desired.OptStartSeq = info.Config.OptStartSeq
-	if _, err = js.UpdatePushConsumer(ctx, stream, desired); err == nil {
+	if _, err = js.UpdatePushConsumer(ctx, lane.stream, desired); err == nil {
 		return desired.DeliverSubject, nil
 	}
 	if !requiresConsumerReplacement(err) {
@@ -438,11 +242,11 @@ func ensureFlowConsumer(nc *nats.Conn, stream string, desired jsapi.ConsumerConf
 	}
 
 	carryLaneAckFloor(&desired, info)
-	return desired.DeliverSubject, replaceFlowConsumer(js, stream, desired, err)
+	return desired.DeliverSubject, replaceFlowConsumer(js, lane, desired, err)
 }
 
-func flowConsumerInfo(ctx context.Context, js jsapi.JetStream, stream, name string) (*jsapi.ConsumerInfo, error) {
-	consumer, err := js.PushConsumer(ctx, stream, name)
+func flowConsumerInfo(ctx context.Context, js jsapi.JetStream, lane laneBinding) (*jsapi.ConsumerInfo, error) {
+	consumer, err := js.PushConsumer(ctx, lane.stream, lane.consumer)
 	if errors.Is(err, jsapi.ErrConsumerNotFound) {
 		return nil, nil
 	}
@@ -462,18 +266,18 @@ func flowConsumerInfo(ctx context.Context, js jsapi.JetStream, stream, name stri
 // desired's delivery position, so the recreation never replays handled messages.
 func replaceFlowConsumer(
 	js jsapi.JetStream,
-	stream string,
+	lane laneBinding,
 	desired jsapi.ConsumerConfig,
 	cause error,
 ) error {
 	ctx, cancel := context.WithTimeout(context.Background(), laneReplaceTimeout)
 	defer cancel()
 
-	if err := js.DeleteConsumer(ctx, stream, desired.Name); err != nil &&
+	if err := js.DeleteConsumer(ctx, lane.stream, desired.Name); err != nil &&
 		!errors.Is(err, jsapi.ErrConsumerNotFound) {
 		return fmt.Errorf("bus: update flow consumer %q: %w (replace failed: %v)", desired.Name, cause, err)
 	}
-	if _, err := js.CreateOrUpdatePushConsumer(ctx, stream, desired); err != nil {
+	if _, err := js.CreateOrUpdatePushConsumer(ctx, lane.stream, desired); err != nil {
 		return fmt.Errorf("bus: recreate flow consumer %q: %w", desired.Name, err)
 	}
 	return nil
@@ -494,660 +298,4 @@ func carryLaneAckFloor(desired *jsapi.ConsumerConfig, info *jsapi.ConsumerInfo) 
 	}
 	desired.DeliverPolicy = jsapi.DeliverByStartSequencePolicy
 	desired.OptStartSeq = info.AckFloor.Stream + 1
-}
-
-// flowPosition is one immutable receipt cursor sample.
-type flowPosition struct {
-	consumer uint64
-	stream   uint64
-}
-
-// flowCursor tracks the highest delivery this process has actually taken off
-// the wire. The flow response never claims more than this, so the replicated
-// ack floor can never move past a message the pod did not receive.
-type flowCursor struct {
-	mu       sync.Mutex
-	position flowPosition
-}
-
-// record gates on the STREAM sequence, which is monotonic across a consumer
-// delete + recreate. The consumer sequence is not: the server restarts it at 1
-// whatever start sequence the recreated consumer opens at, so a cursor gated on
-// it rejects every delivery from the new session forever and freezes the ack
-// floor at MaxAckPending with no redelivery and no error. It returns false for a
-// sample that did not advance the cursor.
-func (c *flowCursor) record(consumer, stream uint64) bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.isSessionReset(consumer) {
-		c.position = flowPosition{consumer: consumer, stream: stream}
-		return true
-	}
-	if stream <= c.position.stream {
-		return false
-	}
-	c.position = flowPosition{consumer: consumer, stream: stream}
-	return true
-}
-
-// isSessionReset recognises the server's own reset signature: a consumer
-// sequence further back than a whole ack window cannot be reordering.
-func (c *flowCursor) isSessionReset(consumer uint64) bool {
-	return c.position.consumer > consumer &&
-		c.position.consumer-consumer > flowSessionResetGap
-}
-
-func (c *flowCursor) snapshot() flowPosition {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	return c.position
-}
-
-// reset drops the cursor when this process recreates its own consumer. The new
-// consumer's sequences start over, so keeping the old pair would answer flow
-// control with an ack floor from a session the server no longer has.
-func (c *flowCursor) reset() {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.position = flowPosition{}
-}
-
-// flowControlResponse builds the receipt-level acknowledgement for one push
-// status message, or nil when the status needs no answer (a heartbeat sent
-// while no flow-control window is outstanding).
-//
-// A response is mandatory even with nothing received: the server reopens the
-// delivery window purely on the response arriving, and only reads an ack floor
-// out of the headers when they are present.
-func flowControlResponse(control *nats.Msg, position flowPosition) *nats.Msg {
-	reply := control.Reply
-	if stalled := control.Header.Get(consumerStalledHeader); stalled != "" {
-		reply = stalled
-	}
-	if reply == "" {
-		return nil
-	}
-	response := nats.NewMsg(reply)
-	if position.stream == 0 {
-		return response
-	}
-	response.Header.Set(lastConsumerHeader, strconv.FormatUint(position.consumer, 10))
-	response.Header.Set(lastStreamHeader, strconv.FormatUint(position.stream, 10))
-	return response
-}
-
-type flowLaneConfig struct {
-	url     string
-	stream  string
-	subject string
-	group   string
-	log     *zap.Logger
-}
-
-// flowDelivery pairs a decoded message with the wire it arrived on, so the
-// retry path can re-send the original bytes and headers without re-encoding.
-type flowDelivery struct {
-	wire *nats.Msg
-	msg  *Message
-}
-
-// flowSubscriber is the receipt-level delivery adapter for the hot ingress
-// lanes. It binds this pod's own AckFlowControl push consumer through a plain
-// core subscription on the consumer's delivery subject: individual messages
-// carry no acknowledgement traffic at all, and the flow-control responses this
-// adapter sends advance the whole window's ack floor in one replicated step.
-//
-// The subscription is deliberately a core one rather than nats.go's own push
-// binding. Both of nats.go's flow-control handlers answer a request with an
-// empty message, and the server only reads an ack floor out of a response that
-// carries the sequence headers — so under AckFlowControl their responses reopen
-// the delivery window without ever acknowledging anything, and the consumer's
-// pending set would grow until MaxAckPending stalled it.
-//
-// One subscriber serves every consumer unit in the pod: the units share the one
-// output channel, so the pod keeps exactly one consumer, one cursor and one
-// flow-control conversation however far the routine autoscaler grows.
-type flowSubscriber struct {
-	nc             *nats.Conn
-	stream         string
-	subject        string
-	group          string
-	consumer       string
-	deliverSubject string
-	log            *zap.Logger
-
-	cursor      flowCursor
-	dropped     atomic.Int64
-	retried     atomic.Int64
-	stale       atomic.Int64
-	overrun     atomic.Int64
-	stallStreak atomic.Int64
-	wedged      atomic.Int64
-	lastControl atomic.Int64
-	recovering  atomic.Bool
-	closed      atomic.Bool
-
-	sub     *nats.Subscription
-	queue   chan flowDelivery
-	output  chan *Message
-	closeCh chan struct{}
-	// ctx is the lane's own delivery context, carried on every message. Units come
-	// and go around the shared binding, so a single unit's context cannot own the
-	// cancellation every handler reads; this one is cancelled when the binding
-	// itself is released.
-	ctx    context.Context
-	cancel context.CancelFunc
-
-	workers sync.WaitGroup
-	pending sync.WaitGroup
-	once    sync.Once
-}
-
-// newFlowLaneSubscriber connects, provisions this pod's flow consumer, and binds
-// the delivery subject the server reports.
-func newFlowLaneSubscriber(cfg flowLaneConfig) (*flowSubscriber, error) {
-	consumer := flowConsumerName(cfg.group, cfg.subject)
-
-	// The connection name separates the two acknowledgement contracts in NATS
-	// monitoring, so a mixed rollout is legible from connz alone.
-	nc, err := nats.Connect(busURL(cfg.url), busOptions(cfg.group+"-flow")...)
-	if err != nil {
-		return nil, err
-	}
-	deliver, err := ensureFlowConsumer(nc, cfg.stream, flowConsumerConfig(cfg.subject, consumer))
-	if err != nil {
-		nc.Close()
-		return nil, err
-	}
-
-	s := newFlowSubscriber(cfg, nc, consumer, deliver)
-	if err := s.start(); err != nil {
-		nc.Close()
-		return nil, err
-	}
-	return s, nil
-}
-
-// There is deliberately no exported constructor that takes a stream and subject
-// directly. One existed for the load rig, and it was the only way past the two
-// guards the service path applies: streamForTopic, which refuses a subject the
-// catalog does not own, and isHotIngressLane, which refuses any stream but the
-// two ingress lane streams. With the rig gone, every caller goes through NewSubscriber and
-// those two keep deciding which lanes get receipt-level acknowledgement.
-
-func newFlowSubscriber(cfg flowLaneConfig, nc *nats.Conn, consumer, deliver string) *flowSubscriber {
-	log := cfg.log
-	if log == nil {
-		log = zap.NewNop()
-	}
-	ctx, cancel := context.WithCancel(context.Background())
-	return &flowSubscriber{
-		nc: nc, stream: cfg.stream, subject: cfg.subject, group: cfg.group,
-		consumer: consumer, deliverSubject: deliver, log: log,
-		// Sized at the server's own byte window, not at MaxAckPending: the window
-		// is what actually stops delivery under this ack policy, and it ramps well
-		// past the message figure.
-		queue:   make(chan flowDelivery, flowQueueDepth),
-		output:  make(chan *Message),
-		closeCh: make(chan struct{}),
-		ctx:     ctx,
-		cancel:  cancel,
-	}
-}
-
-func (s *flowSubscriber) start() error {
-	sub, err := s.nc.Subscribe(s.deliverSubject, s.deliveryCallback)
-	if err != nil {
-		return err
-	}
-	s.sub = sub
-	s.lastControl.Store(time.Now().UnixNano())
-	s.workers.Add(2)
-	go s.pump()
-	go s.watchHeartbeats()
-	return nil
-}
-
-// Subscribe hands back the pod's single lane channel. Every consumer unit reads
-// the same channel, which is what keeps one consumer per pod while the unit
-// autoscaler grows and shrinks.
-func (s *flowSubscriber) Subscribe(_ context.Context, subject string) (<-chan *Message, error) {
-	if subject != s.subject {
-		return nil, fmt.Errorf("bus: flow subscriber is bound to %q, not %q", s.subject, subject)
-	}
-	if s.closed.Load() {
-		return nil, errors.New("bus: subscriber is closed")
-	}
-	return s.output, nil
-}
-
-// deliveryCallback runs on nats.go's single per-subscription goroutine. Status
-// messages are answered here, before anything that can block: a flow-control
-// request queued behind a data message the handler pool has not accepted yet
-// leaves the server's window closed, and delivery stops for the whole consumer
-// until the pool drains. Data is handed to a bounded queue and never waits.
-func (s *flowSubscriber) deliveryCallback(wire *nats.Msg) {
-	if status := wire.Header.Get(statusHeader); status != "" {
-		s.handleStatus(wire, status)
-		return
-	}
-	s.enqueue(wire)
-}
-
-// enqueue records the receipt cursor before queueing, so a flow-control request
-// arriving straight after the delivery can already acknowledge it.
-func (s *flowSubscriber) enqueue(wire *nats.Msg) {
-	// A delivery is the proof the window reopened, so it ends any wedge streak the
-	// stalled heartbeats had accumulated.
-	s.stallStreak.Store(0)
-	s.recordReceipt(wire)
-	msg, err := messageFromNATS(wire)
-	if err != nil {
-		s.log.Warn("dropping malformed lane delivery", zap.String("subject", s.subject), zap.Error(err))
-		return
-	}
-	msg.SetContext(s.laneContext())
-	select {
-	case s.queue <- flowDelivery{wire: wire, msg: msg}:
-	default:
-		// Withholding the flow-control response would stall the consumer; dropping
-		// one delivery costs one event. The queue is cut to sit above the server's
-		// ramped flow-control window, so this stays an alert rather than a mode of
-		// operation — a sustained count here means the window grew past what
-		// flowQueueDepth was derived for and the derivation needs re-measuring.
-		//
-		// Throttled like the stale counter, and for the same reason: sustained
-		// overflow is exactly the case where a line per drop would emit at the
-		// event rate of the lane. The counter still moves on every drop, so the
-		// magnitude is never lost — only the repetition is.
-		if overrun := s.overrun.Add(1); overrun == 1 || overrun%1_000 == 0 {
-			s.log.Warn("lane receipt queue overflowed",
-				zap.String("subject", s.subject),
-				zap.Int64("overrun_total", overrun))
-		}
-	}
-}
-
-// laneContext keeps a directly-constructed subscriber (tests) usable without a
-// context of its own.
-func (s *flowSubscriber) laneContext() context.Context {
-	if s.ctx == nil {
-		return context.Background()
-	}
-	return s.ctx
-}
-
-func (s *flowSubscriber) recordReceipt(wire *nats.Msg) {
-	metadata, err := wire.Metadata()
-	if err != nil {
-		return
-	}
-	if s.cursor.record(metadata.Sequence.Consumer, metadata.Sequence.Stream) {
-		return
-	}
-	// Deliveries are arriving that the cursor refuses to account for, so the ack
-	// floor has stopped advancing while the pending set keeps growing.
-	if stale := s.stale.Add(1); stale == 1 || stale%10_000 == 0 {
-		s.log.Warn("receipt cursor is not advancing under delivery",
-			zap.String("subject", s.subject),
-			zap.Uint64("delivery_stream_seq", metadata.Sequence.Stream),
-			zap.Int64("stale_total", stale))
-	}
-}
-
-// pump moves queued deliveries onto the shared lane channel. It is the only
-// sender, so the channel can be closed once it has exited.
-func (s *flowSubscriber) pump() {
-	defer s.workers.Done()
-	for {
-		select {
-		case delivery := <-s.queue:
-			if !s.deliver(delivery) {
-				return
-			}
-		case <-s.closeCh:
-			return
-		}
-	}
-}
-
-// deliver hands one message to the lane channel with its verdict already wired
-// up. The reconciliation is a callback rather than a goroutine parked on
-// Acked/Nacked, because at lane rate that goroutine and its 30-second timer were
-// a spawn and a timer-heap insert per message for an outcome that, in the common
-// case, does nothing at all: an ACK is receipt-level here and needs no traffic,
-// since the flow-control responses already own the ack floor. Only a NACK costs
-// anything, and it costs one scheduled retry.
-//
-// Two consequences are deliberate. The retry now runs on the handler's own
-// goroutine, bounded by flowRetryTimeout, which is affordable because failure is
-// rare by contract on these lanes. And a handler that resolves neither way holds
-// its pending count until shutdown, where subscriberDrainTimeout bounds the wait
-// — the same 30 seconds the per-message timer used to spend, now paid once at
-// shutdown instead of being armed on every delivery.
-func (s *flowSubscriber) deliver(delivery flowDelivery) bool {
-	// Both the count and the handler are in place before the message is visible
-	// to a handler: it can be resolved the instant the send completes.
-	s.pending.Add(1)
-	delivery.msg.setResolveHandler(func(acked bool) {
-		defer s.pending.Done()
-		if !acked {
-			s.scheduleRetry(delivery)
-		}
-	})
-	select {
-	case s.output <- delivery.msg:
-		return true
-	case <-s.closeCh:
-		// Nobody took the message, so the callback will never run and the count
-		// it owns has to be released here or shutdown waits out its whole budget.
-		s.pending.Done()
-		return false
-	}
-}
-
-func (s *flowSubscriber) scheduleRetry(delivery flowDelivery) {
-	if err := scheduleLaneRetry(s.nc, s.subject, delivery.wire, delivery.msg); err != nil {
-		s.dropRetry(delivery.msg, err)
-		return
-	}
-	s.retried.Add(1)
-}
-
-// scheduleLaneRetry is the redelivery both receipt-level lane adapters share.
-// Neither has per-message pending state to NAK against: the flow lane's server
-// does not even subscribe to $JS.ACK for its ack policy, and the pull lane's
-// cumulative floor has already moved past the message. Both therefore hand the
-// event to the broker's own scheduler instead — a one-shot @at row on the retry
-// stream that re-emits the payload onto the retry lane once the delay has
-// passed, then purges itself.
-//
-// It is one function on purpose: the budget check, the schedule shape and the
-// PubAck-error reading are the parts that go silently wrong, and a second copy
-// of them would be a second place for a retry to be dropped without a trace.
-func scheduleLaneRetry(nc *nats.Conn, lane string, wire *nats.Msg, msg *Message) error {
-	if attempt := msg.Metadata.Get(RetryCountHeader); attempt != "" {
-		return fmt.Errorf("one-hop retry budget exhausted at attempt %s", attempt)
-	}
-	ack, err := nc.RequestMsg(retryScheduleMsg(lane, wire, flowRetryDelay(), time.Now()), flowRetryTimeout)
-	if err != nil {
-		return err
-	}
-	return pubAckError(ack)
-}
-
-// retryScheduleMsg builds the one-shot schedule row. The row's own subject is
-// unique per retry because publishing a schedule rolls its subject up: a second
-// retry on a shared subject would purge the first one before it ever fired.
-//
-// No Nats-Schedule-Time-Zone header is set. The server rejects the @at form
-// outright when a time zone is present, even "UTC".
-func retryScheduleMsg(lane string, wire *nats.Msg, delay time.Duration, now time.Time) *nats.Msg {
-	target := RetryLaneSubject(lane)
-	schedule := nats.NewMsg(target + "." + nuid.Next())
-	schedule.Data = append([]byte(nil), wire.Data...)
-	copyApplicationHeaders(schedule.Header, wire.Header)
-
-	// Carry a stable idempotency key across the hop. An ingress-origin event
-	// reaches the lane with no Bagelbot-Message-Id, so copyApplicationHeaders had
-	// nothing to carry and the re-emitted retry would land under a fresh stream
-	// sequence — a different Message.UUID that a consumer's dedup guard could not
-	// recognise as the same event. Stamp the delivery's already-resolved identity
-	// (the JetStream stream-seq fallback for those events) so Message.UUID stays
-	// stable across the retry. A Go-origin header copied above is left untouched.
-	if wire.Header.Get(MessageIDHeader) == "" {
-		schedule.Header.Set(MessageIDHeader, messageIdentity(wire))
-	}
-
-	schedule.Header.Set(jsapi.ScheduleHeader, "@at "+now.Add(delay).UTC().Format(time.RFC3339))
-	schedule.Header.Set(jsapi.ScheduleTargetHeader, target)
-	schedule.Header.Set(jsapi.ScheduleTTLHeader, retryScheduleTTL)
-	schedule.Header.Set(RetryCountHeader, "1")
-	return schedule
-}
-
-// copyApplicationHeaders carries the fleet's own headers across the hop and
-// leaves every Nats-* header behind. The scheduler strips those from what it
-// emits anyway, and the expectation headers among them would be re-evaluated
-// against the retry stream and reject the publish.
-func copyApplicationHeaders(into, from nats.Header) {
-	for key, values := range from {
-		if len(values) == 0 || strings.HasPrefix(key, "Nats-") {
-			continue
-		}
-		into.Set(key, values[0])
-	}
-}
-
-// flowRetryDelay is how long a failed event waits before the broker re-emits it.
-// It exists to let a transient downstream failure clear, so it is tunable
-// without a rebuild.
-func flowRetryDelay() time.Duration {
-	delay := env.GetDuration("NATS_FLOW_RETRY_DELAY", defaultFlowRetryDelay)
-	if delay <= 0 {
-		return defaultFlowRetryDelay
-	}
-	return delay
-}
-
-// pubAckError reports a JetStream publish rejection. A schedule the broker
-// refuses (schedules disabled, an invalid target, a denied subject) answers with
-// a normal PubAck carrying an error, which is silent unless it is read.
-func pubAckError(ack *nats.Msg) error {
-	if ack == nil {
-		return errors.New("bus: retry schedule got no publish acknowledgement")
-	}
-	var response struct {
-		Error *struct {
-			ErrCode     int    `json:"err_code"`
-			Description string `json:"description"`
-		} `json:"error"`
-	}
-	if err := sonic.Unmarshal(ack.Data, &response); err != nil {
-		return fmt.Errorf("bus: unreadable retry publish acknowledgement: %w", err)
-	}
-	if response.Error != nil {
-		return fmt.Errorf("bus: retry schedule rejected (%d): %s",
-			response.Error.ErrCode, response.Error.Description)
-	}
-	return nil
-}
-
-func (s *flowSubscriber) dropRetry(msg *Message, cause error) {
-	s.log.Warn("dropping failed lane event",
-		zap.String("subject", s.subject),
-		zap.String("message_id", msg.UUID),
-		zap.Int64("dropped_total", s.dropped.Add(1)),
-		zap.Error(cause))
-}
-
-// handleStatus answers control messages and records that the consumer behind
-// this delivery subject is alive. Nothing but 100-class control messages is ever
-// published to a push delivery subject in 2.14.3 — every 404 and 409 goes to a
-// pull request's reply — so a non-100 status is logged rather than treated as a
-// deletion signal. Deletion is detected by the heartbeat watchdog instead.
-func (s *flowSubscriber) handleStatus(control *nats.Msg, status string) {
-	if status == controlStatus {
-		s.lastControl.Store(time.Now().UnixNano())
-		s.noteStall(control)
-		s.noteDeliveryGap(control)
-		s.respondFlowControl(control)
-		return
-	}
-	s.log.Warn("unexpected flow consumer status",
-		zap.String("subject", s.subject),
-		zap.String("status", status),
-		zap.String("description", control.Header.Get(descriptionHeader)))
-}
-
-// noteStall counts stalled heartbeats that no delivery interrupted. The stalled
-// marker says the server is holding the window shut until a flow-control
-// response reaches it, so a run of them says this pod's responses are not
-// arriving — and the only two causes are an ACL that denies publishing to the
-// $JS.FC subject and a response the connection dropped.
-//
-// It is reported, never repaired. Recreating the consumer cannot grant a
-// permission, and a watchdog that recreated on this signal would churn the lane
-// once a second for as long as the denial lasted; the ack floor and every
-// in-flight delivery would go with it each time. The counter and the log line
-// are what an operator needs, and the fix is on the broker's side.
-func (s *flowSubscriber) noteStall(control *nats.Msg) {
-	fcid := control.Header.Get(consumerStalledHeader)
-	if fcid == "" {
-		return
-	}
-	if streak := s.stallStreak.Add(1); streak%flowWedgeStreak == 0 {
-		s.log.Error("flow consumer is wedged: stalled heartbeats are going unanswered",
-			zap.String("subject", s.subject),
-			zap.String("consumer", s.consumer),
-			zap.String("flow_control_id", fcid),
-			zap.Int64("stalled_heartbeats", streak),
-			zap.Int64("wedged_total", s.wedged.Add(1)))
-	}
-}
-
-// noteDeliveryGap compares what the heartbeat says the server has sent with what
-// this process has taken off the wire. A gap wider than the whole ack window
-// cannot be traffic in flight: after a leader change the server resumes past its
-// last delivered sequence and never redelivers what fell in between, and this ack
-// policy has no other redelivery. It is reported rather than repaired —
-// $JS.API.CONSUMER.RESET would rewind to the ack floor and replay everything
-// already handled since it.
-func (s *flowSubscriber) noteDeliveryGap(control *nats.Msg) {
-	sent, err := strconv.ParseUint(control.Header.Get(lastConsumerHeader), 10, 64)
-	if err != nil {
-		return
-	}
-	received := s.cursor.snapshot().consumer
-	if !deliveriesWereLost(sent, received) {
-		return
-	}
-	s.log.Warn("flow consumer deliveries were never received",
-		zap.String("subject", s.subject),
-		zap.Uint64("server_last_consumer", sent),
-		zap.Uint64("received_consumer", received))
-}
-
-// deliveriesWereLost separates a legitimate in-flight window from a real gap.
-// The server never has more than MaxAckPending deliveries outstanding, so
-// anything beyond that is traffic that will never arrive.
-func deliveriesWereLost(sent, received uint64) bool {
-	return sent > received+flowMaxAckPending
-}
-
-func (s *flowSubscriber) respondFlowControl(control *nats.Msg) {
-	response := flowControlResponse(control, s.cursor.snapshot())
-	if response == nil {
-		return
-	}
-	if err := s.nc.PublishMsg(response); err != nil {
-		// The server reopens the window on its next heartbeat's stalled marker,
-		// so a lost response costs latency rather than the binding.
-		s.log.Warn("flow-control response failed", zap.String("subject", s.subject), zap.Error(err))
-	}
-}
-
-// watchHeartbeats is the only way this binding learns its consumer is gone. The
-// server pins the heartbeat at one second and publishes no deletion status to a
-// push subject, so silence on the delivery subject is the signal: the lane would
-// otherwise sit at zero events per second, healthy and Ready, indefinitely.
-//
-// It detects a DELETED consumer and nothing else. A wedged one heartbeats
-// forever, so silence never comes and this watchdog never fires for it; that
-// failure belongs to noteStall.
-func (s *flowSubscriber) watchHeartbeats() {
-	defer s.workers.Done()
-	ticker := time.NewTicker(flowWatchdogTick)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-s.closeCh:
-			return
-		case now := <-ticker.C:
-			if heartbeatLost(now, s.lastControl.Load()) {
-				s.recoverConsumer()
-			}
-		}
-	}
-}
-
-// heartbeatLost is the watchdog's decision. The heartbeat is pinned at one
-// second for this ack policy, so three missed beats is silence, not jitter.
-func heartbeatLost(now time.Time, lastControlUnixNano int64) bool {
-	return now.Sub(time.Unix(0, lastControlUnixNano)) > flowHeartbeatGrace
-}
-
-// recoverConsumer re-provisions a consumer the watchdog found silent. The
-// delivery subject is deterministic, so the existing subscription keeps
-// receiving once the consumer is back; single-flighting stops the watchdog from
-// racing itself while a slow provision is in progress.
-func (s *flowSubscriber) recoverConsumer() {
-	if !s.recovering.CompareAndSwap(false, true) {
-		return
-	}
-	s.pending.Add(1)
-	go func() {
-		defer s.pending.Done()
-		defer s.recovering.Store(false)
-		s.reprovision()
-	}()
-}
-
-func (s *flowSubscriber) reprovision() {
-	desired := recoveryFlowConsumerConfig(s.subject, s.consumer, s.cursor.snapshot())
-	if _, err := ensureFlowConsumer(s.nc, s.stream, desired); err != nil {
-		s.log.Error("flow consumer recovery failed", zap.String("subject", s.subject), zap.Error(err))
-		return
-	}
-	// The recreated consumer restarts its own sequences, so the old pair would
-	// answer flow control for a session the server no longer has.
-	s.cursor.reset()
-	s.lastControl.Store(time.Now().UnixNano())
-	s.log.Warn("flow consumer re-provisioned after heartbeat loss",
-		zap.String("subject", s.subject), zap.String("consumer", s.consumer))
-}
-
-func (s *flowSubscriber) Close() error {
-	var err error
-	s.once.Do(func() { err = s.shutdown() })
-	return err
-}
-
-func (s *flowSubscriber) shutdown() error {
-	s.closed.Store(true)
-	if s.cancel != nil {
-		s.cancel()
-	}
-	if s.sub != nil {
-		if unsub := s.sub.Unsubscribe(); unsub != nil && !errors.Is(unsub, nats.ErrBadSubscription) {
-			s.log.Warn("flow subscription stop failed", zap.String("subject", s.subject), zap.Error(unsub))
-		}
-	}
-	close(s.closeCh)
-	s.workers.Wait()
-
-	deadline := time.NewTimer(subscriberDrainTimeout)
-	defer deadline.Stop()
-	drained := waitGroupBefore(&s.pending, deadline.C)
-
-	err := s.flushAndClose(drained)
-	close(s.output)
-	return err
-}
-
-// flushAndClose pushes the last flow-control response and retry schedule onto
-// the wire before the connection goes away. Closing without a flush discards
-// whatever is still buffered, which on this lane is exactly the acknowledgement
-// that would have released the final window.
-func (s *flowSubscriber) flushAndClose(drained bool) error {
-	flushErr := s.nc.FlushTimeout(2 * time.Second)
-	s.nc.Close()
-	if !drained {
-		return errors.New("bus: timed out draining flow lane deliveries")
-	}
-	if flushErr != nil {
-		return fmt.Errorf("bus: flush flow lane acknowledgements: %w", flushErr)
-	}
-	return nil
 }
