+++ pkg/bus/flow_consumer_test.go
@@ -9,7 +9,6 @@ import (
 
 	"github.com/nats-io/nats.go"
 	jsapi "github.com/nats-io/nats.go/jetstream"
-	"go.uber.org/zap"
 )
 
 func TestOnlyHotIngressLanesUseFlowControl(t *testing.T) {
@@ -38,7 +37,7 @@ func TestOnlyHotIngressLanesUseFlowControl(t *testing.T) {
 }
 
 func TestFlowConsumerConfigIsAcceptedByTheServerContract(t *testing.T) {
-	cfg := flowConsumerConfig("twitch.ingress.event.premium", "worker_premium_pod_1")
+	cfg := flowConsumerConfig(laneBinding{subject: "twitch.ingress.event.premium", consumer: "worker_premium_pod_1"})
 
 	if cfg.AckPolicy != jsapi.AckFlowControlPolicy {
 		t.Fatalf("ack policy = %v, want AckFlowControl", cfg.AckPolicy)
@@ -67,7 +66,7 @@ func TestFlowConsumerConfigIsAcceptedByTheServerContract(t *testing.T) {
 func TestFlowConsumerIsSingleSubscriberPerPod(t *testing.T) {
 	t.Setenv("POD_NAME", "sesame-6d9f7c8b45-tq2xz")
 	name := flowConsumerName("worker", "twitch.ingress.event.premium")
-	cfg := flowConsumerConfig("twitch.ingress.event.premium", name)
+	cfg := flowConsumerConfig(laneBinding{subject: "twitch.ingress.event.premium", consumer: name})
 
 	// A queue group would hand one arbitrary member every flow-control request,
 	// and its cumulative answer would acknowledge work still running elsewhere.
@@ -103,21 +102,21 @@ func TestPodIdentityFallsBackAndStaysNameSafe(t *testing.T) {
 
 func TestFlowConsumerRecoveryResumesAfterTheReceiptCursor(t *testing.T) {
 	cfg := recoveryFlowConsumerConfig(
-		"twitch.ingress.event.standard", "worker_standard_pod_1",
+		laneBinding{subject: "twitch.ingress.event.standard", consumer: "worker_standard_pod_1"},
 		flowPosition{consumer: 90, stream: 120},
 	)
 	if cfg.DeliverPolicy != jsapi.DeliverByStartSequencePolicy || cfg.OptStartSeq != 121 {
 		t.Fatalf("recovery policy=%v start=%d, want start sequence 121", cfg.DeliverPolicy, cfg.OptStartSeq)
 	}
 
-	fresh := recoveryFlowConsumerConfig("twitch.ingress.event.standard", "worker_standard_pod_1", flowPosition{})
+	fresh := recoveryFlowConsumerConfig(laneBinding{subject: "twitch.ingress.event.standard", consumer: "worker_standard_pod_1"}, flowPosition{})
 	if fresh.DeliverPolicy != jsapi.DeliverNewPolicy {
 		t.Fatalf("empty cursor recovery policy = %v, want DeliverNew", fresh.DeliverPolicy)
 	}
 }
 
 func TestCarriedAckFloorSurvivesConsumerReplacement(t *testing.T) {
-	desired := flowConsumerConfig("twitch.ingress.event.premium", "worker_premium_pod_1")
+	desired := flowConsumerConfig(laneBinding{subject: "twitch.ingress.event.premium", consumer: "worker_premium_pod_1"})
 	info := &jsapi.ConsumerInfo{}
 	info.AckFloor.Stream = 4_200
 
@@ -128,7 +127,7 @@ func TestCarriedAckFloorSurvivesConsumerReplacement(t *testing.T) {
 }
 
 func TestUnknownAckFloorNeverInheritsThePredecessorPolicy(t *testing.T) {
-	desired := flowConsumerConfig("twitch.ingress.event.premium", "worker_premium_pod_1")
+	desired := flowConsumerConfig(laneBinding{subject: "twitch.ingress.event.premium", consumer: "worker_premium_pod_1"})
 	// The explicit-ACK consumer this replaces is DeliverAll; inheriting it would
 	// open the replacement on the whole retained firehose.
 	desired.DeliverPolicy = jsapi.DeliverAllPolicy
@@ -166,478 +165,6 @@ func TestOnlyImmutableFieldErrorsReplaceTheConsumer(t *testing.T) {
 		if got := requiresConsumerReplacement(test.err); got != test.want {
 			t.Fatalf("requiresConsumerReplacement(%v) = %v, want %v", test.err, got, test.want)
 		}
-	}
-}
-
-func TestFlowControlResponseCarriesTheReceiptCursor(t *testing.T) {
-	control := flowStatus("FlowControl Request")
-	control.Reply = "$JS.FC.TWITCH_INGRESS.worker_premium.abcd"
-
-	response := flowControlResponse(control, flowPosition{consumer: 42, stream: 99})
-	if response == nil {
-		t.Fatal("flow-control request produced no response")
-	}
-	if response.Subject != control.Reply {
-		t.Fatalf("response subject = %q, want %q", response.Subject, control.Reply)
-	}
-	if response.Header.Get(lastConsumerHeader) != "42" || response.Header.Get(lastStreamHeader) != "99" {
-		t.Fatalf("response cursor = %#v", response.Header)
-	}
-}
-
-func TestIdleHeartbeatIsAnsweredOnlyWhileStalled(t *testing.T) {
-	quiet := flowStatus("Idle Heartbeat")
-	quiet.Header.Set(lastConsumerHeader, "500")
-	quiet.Header.Set(lastStreamHeader, "900")
-	if response := flowControlResponse(quiet, flowPosition{consumer: 42, stream: 99}); response != nil {
-		t.Fatalf("unstalled heartbeat produced response %#v", response)
-	}
-
-	stalled := flowStatus("Idle Heartbeat")
-	stalled.Header.Set(consumerStalledHeader, "$JS.FC.TWITCH_INGRESS.worker_premium.stall")
-	stalled.Header.Set(lastConsumerHeader, "500")
-	stalled.Header.Set(lastStreamHeader, "900")
-
-	response := flowControlResponse(stalled, flowPosition{consumer: 42, stream: 99})
-	if response == nil || response.Subject != "$JS.FC.TWITCH_INGRESS.worker_premium.stall" {
-		t.Fatalf("stalled heartbeat response = %#v", response)
-	}
-	// The server reports what it sent, not what arrived. Answering with the
-	// server's cursor would advance the replicated ack floor past deliveries
-	// this process never received.
-	if response.Header.Get(lastConsumerHeader) != "42" || response.Header.Get(lastStreamHeader) != "99" {
-		t.Fatalf("stalled response used the server cursor: %#v", response.Header)
-	}
-}
-
-func TestFlowControlIsAnsweredBeforeAnyDelivery(t *testing.T) {
-	control := flowStatus("FlowControl Request")
-	control.Reply = "$JS.FC.TWITCH_INGRESS.worker_premium.abcd"
-
-	// The window only reopens on the response arriving, so an empty cursor must
-	// still answer — it simply claims no ack floor.
-	response := flowControlResponse(control, flowPosition{})
-	if response == nil {
-		t.Fatal("empty cursor produced no response")
-	}
-	if response.Header.Get(lastStreamHeader) != "" {
-		t.Fatalf("empty cursor claimed an ack floor: %#v", response.Header)
-	}
-}
-
-func TestReceiptCursorTracksTheStreamSequence(t *testing.T) {
-	cursor := &flowCursor{}
-	if !cursor.record(10, 100) {
-		t.Fatal("first sample was refused")
-	}
-	// Stale samples inside the same session never move the floor backwards.
-	if cursor.record(4, 40) {
-		t.Fatal("a stale sample advanced the cursor")
-	}
-	if got := cursor.snapshot(); got.consumer != 10 || got.stream != 100 {
-		t.Fatalf("cursor regressed to %+v", got)
-	}
-	if !cursor.record(11, 101) || cursor.snapshot().stream != 101 {
-		t.Fatalf("cursor did not advance: %+v", cursor.snapshot())
-	}
-}
-
-func TestReceiptCursorSurvivesAConsumerSequenceReset(t *testing.T) {
-	cursor := &flowCursor{}
-	cursor.record(5_000_000, 9_000_000)
-
-	// A delete + recreate restarts the consumer sequence at 1 while the stream
-	// sequence keeps climbing. Gating on the consumer sequence would reject every
-	// delivery of the new session forever.
-	if !cursor.record(1, 9_000_001) {
-		t.Fatal("the recreated consumer's first delivery was refused")
-	}
-	if got := cursor.snapshot(); got.consumer != 1 || got.stream != 9_000_001 {
-		t.Fatalf("cursor after reset = %+v", got)
-	}
-
-	// A recreation at an earlier start sequence moves both coordinates back: the
-	// backwards jump in consumer sequence is the server's own reset signature.
-	rewound := &flowCursor{}
-	rewound.record(5_000_000, 9_000_000)
-	if !rewound.record(2, 12) {
-		t.Fatal("a session reset below the stored stream sequence was refused")
-	}
-	if got := rewound.snapshot(); got.stream != 12 {
-		t.Fatalf("cursor after a rewound session = %+v", got)
-	}
-}
-
-func TestReceiptCursorResetsOnSelfInitiatedRecreate(t *testing.T) {
-	cursor := &flowCursor{}
-	cursor.record(90, 120)
-	cursor.reset()
-	if got := cursor.snapshot(); got != (flowPosition{}) {
-		t.Fatalf("cursor after reset = %+v, want empty", got)
-	}
-}
-
-func TestRetryScheduleMatchesTheServerScheduleContract(t *testing.T) {
-	wire := laneDelivery("logical-id", []byte(`{"text":"hello"}`))
-	wire.Header.Set("Nats-Expected-Last-Sequence", "9")
-	wire.Header.Set("traceparent", "00-trace-span-01")
-	now := time.Date(2026, 7, 27, 12, 0, 0, 0, time.UTC)
-
-	schedule := retryScheduleMsg("twitch.ingress.event.premium", wire, 3*time.Second, now)
-
-	if got := schedule.Header.Get(jsapi.ScheduleHeader); got != "@at 2026-07-27T12:00:03Z" {
-		t.Fatalf("schedule pattern = %q, want a one-shot @at three seconds out", got)
-	}
-	if got := schedule.Header.Get(jsapi.ScheduleTargetHeader); got != "twitch.ingress.retry.premium" {
-		t.Fatalf("schedule target = %q", got)
-	}
-	// The server rejects @at outright when a time zone header is present, even UTC.
-	if schedule.Header.Get(jsapi.ScheduleTimeZoneHeader) != "" {
-		t.Fatalf("schedule carried a time zone: %#v", schedule.Header)
-	}
-	if schedule.Header.Get(jsapi.ScheduleTTLHeader) != retryScheduleTTL {
-		t.Fatalf("schedule TTL = %q", schedule.Header.Get(jsapi.ScheduleTTLHeader))
-	}
-	if schedule.Header.Get(RetryCountHeader) != "1" {
-		t.Fatalf("retry count = %q, want the first hop", schedule.Header.Get(RetryCountHeader))
-	}
-	// Application headers survive the emit; Nats-* headers do not, and the
-	// expectation headers among them would be re-evaluated against the retry
-	// stream and reject the publish.
-	if schedule.Header.Get("traceparent") != "00-trace-span-01" {
-		t.Fatalf("application headers were dropped: %#v", schedule.Header)
-	}
-	if schedule.Header.Get(MessageIDHeader) != "logical-id" {
-		t.Fatalf("fleet identity was dropped: %#v", schedule.Header)
-	}
-	if schedule.Header.Get("Nats-Expected-Last-Sequence") != "" {
-		t.Fatalf("an expectation header rode the retry: %#v", schedule.Header)
-	}
-	if string(schedule.Data) != `{"text":"hello"}` {
-		t.Fatalf("retry payload = %q", schedule.Data)
-	}
-}
-
-func TestRetryStampsIdentityWhenIngressWireHasNone(t *testing.T) {
-	// Ingress-origin events reach the lane with no Bagelbot-Message-Id, so the
-	// re-emitted retry would otherwise land under a fresh stream sequence and read
-	// as a different message. The retry must stamp the delivery's resolved
-	// identity so a consumer's dedup guard recognises it as the same event.
-	wire := nats.NewMsg("twitch.ingress.event.standard")
-	wire.Data = []byte(`{"text":"hello"}`)
-	if wire.Header.Get(MessageIDHeader) != "" {
-		t.Fatal("precondition: an ingress wire carries no message id")
-	}
-
-	schedule := retryScheduleMsg("twitch.ingress.event.standard", wire, time.Second, time.Now())
-	if schedule.Header.Get(MessageIDHeader) == "" {
-		t.Fatalf("retry left the fleet identity empty: %#v", schedule.Header)
-	}
-}
-
-func TestRetrySchedulesNeverShareASubject(t *testing.T) {
-	wire := laneDelivery("logical-id", []byte(`{"text":"hello"}`))
-	first := retryScheduleMsg("twitch.ingress.event.standard", wire, time.Second, time.Now())
-	second := retryScheduleMsg("twitch.ingress.event.standard", wire, time.Second, time.Now())
-
-	// Publishing a schedule rolls its subject up, so a shared subject would purge
-	// the previous retry before it ever fired.
-	if first.Subject == second.Subject {
-		t.Fatalf("two retries shared the schedule subject %q", first.Subject)
-	}
-	for _, subject := range []string{first.Subject, second.Subject} {
-		if !strings.HasPrefix(subject, "twitch.ingress.retry.standard.") {
-			t.Fatalf("schedule subject %q left the retry lane's namespace", subject)
-		}
-	}
-}
-
-func TestRetryLaneSubjectIsTheScheduleTarget(t *testing.T) {
-	if got := RetryLaneSubject("twitch.ingress.event.premium"); got != "twitch.ingress.retry.premium" {
-		t.Fatalf("retry lane = %q", got)
-	}
-}
-
-func TestRetryDelayIsTunable(t *testing.T) {
-	t.Setenv("NATS_FLOW_RETRY_DELAY", "")
-	if got := flowRetryDelay(); got != defaultFlowRetryDelay {
-		t.Fatalf("default retry delay = %v", got)
-	}
-	t.Setenv("NATS_FLOW_RETRY_DELAY", "12s")
-	if got := flowRetryDelay(); got != 12*time.Second {
-		t.Fatalf("configured retry delay = %v", got)
-	}
-	t.Setenv("NATS_FLOW_RETRY_DELAY", "-1s")
-	if got := flowRetryDelay(); got != defaultFlowRetryDelay {
-		t.Fatalf("negative retry delay = %v, want the default", got)
-	}
-}
-
-func TestRejectedRetryScheduleIsAnError(t *testing.T) {
-	rejected := nats.NewMsg("reply")
-	rejected.Data = []byte(`{"error":{"code":400,"err_code":10188,"description":"message schedules is disabled"}}`)
-	err := pubAckError(rejected)
-	if err == nil || !strings.Contains(err.Error(), "10188") {
-		t.Fatalf("rejected schedule error = %v", err)
-	}
-
-	accepted := nats.NewMsg("reply")
-	accepted.Data = []byte(`{"stream":"TWITCH_INGRESS_RETRY","seq":12}`)
-	if err := pubAckError(accepted); err != nil {
-		t.Fatalf("accepted schedule reported %v", err)
-	}
-	if pubAckError(nil) == nil {
-		t.Fatal("a missing acknowledgement was accepted")
-	}
-}
-
-func TestAMarkedRetryIsDroppedWithACounter(t *testing.T) {
-	sub := testFlowSubscriber()
-	wire := laneDelivery("logical-id", []byte(`{"text":"hello"}`))
-	wire.Header.Set(RetryCountHeader, "1")
-	msg := mustFlowMessage(t, wire)
-
-	// The budget check runs before anything touches the connection.
-	sub.scheduleRetry(flowDelivery{wire: wire, msg: msg})
-	if sub.dropped.Load() != 1 || sub.retried.Load() != 0 {
-		t.Fatalf("dropped=%d retried=%d, want one drop", sub.dropped.Load(), sub.retried.Load())
-	}
-}
-
-func TestAckIsReceiptLevelAndCostsNothing(t *testing.T) {
-	sub := testFlowSubscriber()
-	defer close(sub.closeCh)
-	msg := deliverToLane(t, sub, laneDelivery("logical-id", []byte(`{"text":"hello"}`)))
-
-	msg.Ack()
-
-	// The resolve callback runs on this goroutine, so the count is already back
-	// by the time Ack returns; Wait is the assertion that it was released at all.
-	sub.pending.Wait()
-	if sub.retried.Load() != 0 || sub.dropped.Load() != 0 {
-		t.Fatalf("a receipt-level ack produced traffic: %d retries, %d drops",
-			sub.retried.Load(), sub.dropped.Load())
-	}
-}
-
-// TestNackRunsTheRetryPathExactlyOnce is the other half of the resolve contract.
-// The verdict no longer has a goroutine of its own, so a callback invoked twice
-// would schedule the event twice and one invoked never would leak the pending
-// count for the whole drain window.
-func TestNackRunsTheRetryPathExactlyOnce(t *testing.T) {
-	sub := testFlowSubscriber()
-	defer close(sub.closeCh)
-	wire := laneDelivery("logical-id", []byte(`{"text":"hello"}`))
-	// An exhausted budget refuses inside the shared helper, before it reaches the
-	// connection, which is what lets the retry path run here without a broker.
-	wire.Header.Set(RetryCountHeader, "1")
-	msg := deliverToLane(t, sub, wire)
-
-	if !msg.Nack() {
-		t.Fatal("the first Nack must win")
-	}
-	// Losing calls must not re-enter the callback.
-	msg.Nack()
-	msg.Ack()
-
-	sub.pending.Wait()
-	if sub.dropped.Load() != 1 || sub.retried.Load() != 0 {
-		t.Fatalf("dropped=%d retried=%d, want exactly one retry attempt",
-			sub.dropped.Load(), sub.retried.Load())
-	}
-}
-
-// TestADeliveryLostToShutdownReleasesItsPendingCount guards the one path where
-// the resolve callback provably never runs: the send lost the race to closeCh,
-// so no handler ever saw the message. Without the explicit release the count
-// would sit there and shutdown would burn its whole drain budget on it.
-func TestADeliveryLostToShutdownReleasesItsPendingCount(t *testing.T) {
-	sub := testFlowSubscriber()
-	wire := laneDelivery("logical-id", []byte(`{"text":"hello"}`))
-	msg := mustFlowMessage(t, wire)
-
-	// Nothing reads the lane channel and the binding is already closing.
-	close(sub.closeCh)
-	if sub.deliver(flowDelivery{wire: wire, msg: msg}) {
-		t.Fatal("a delivery was accepted after the binding closed")
-	}
-
-	deadline := time.NewTimer(2 * time.Second)
-	defer deadline.Stop()
-	if !waitGroupBefore(&sub.pending, deadline.C) {
-		t.Fatal("a delivery lost to shutdown leaked its pending count")
-	}
-}
-
-// deliverToLane runs the real deliver path with a reader standing in for a
-// consumer unit, and hands back the message that unit received.
-func deliverToLane(t *testing.T, sub *flowSubscriber, wire *nats.Msg) *Message {
-	t.Helper()
-	received := make(chan *Message, 1)
-	go func() { received <- <-sub.output }()
-
-	if !sub.deliver(flowDelivery{wire: wire, msg: mustFlowMessage(t, wire)}) {
-		t.Fatal("delivery was refused while the binding was open")
-	}
-	select {
-	case msg := <-received:
-		return msg
-	case <-time.After(2 * time.Second):
-		t.Fatal("the lane never handed the message out")
-		return nil
-	}
-}
-
-func TestStatusHandlingNeverWaitsOnTheDataPath(t *testing.T) {
-	sub := testFlowSubscriber()
-	// One slot, no pump: the data path is as blocked as a saturated routine pool.
-	sub.queue = make(chan flowDelivery, 1)
-
-	done := make(chan struct{})
-	go func() {
-		defer close(done)
-		for i := 0; i < 4; i++ {
-			sub.deliveryCallback(laneDelivery("logical-id", []byte(`{"text":"hello"}`)))
-		}
-	}()
-	select {
-	case <-done:
-	case <-time.After(time.Second):
-		t.Fatal("the delivery callback blocked behind the data path")
-	}
-	if sub.overrun.Load() != 3 {
-		t.Fatalf("overrun counter = %d, want the three refused deliveries", sub.overrun.Load())
-	}
-}
-
-// The queue must clear the server's own brake, and that brake is the byte
-// window, not MaxAckPending: 32 MiB of ~865 B deliveries ramps to ~38k messages
-// in flight, so a queue cut at 20k drops deliveries during healthy delivery.
-func TestReceiptQueueIsSizedOffTheFlowControlByteWindow(t *testing.T) {
-	if flowQueueDepth != flowControlPendingBytes/flowWireBytesFloor {
-		t.Fatalf("queue depth %d is no longer derived from the byte window", flowQueueDepth)
-	}
-	ramped := flowControlPendingBytes / 865
-	if flowQueueDepth <= ramped {
-		t.Fatalf("queue depth %d does not clear the ramped window of %d deliveries", flowQueueDepth, ramped)
-	}
-	if flowQueueDepth <= flowMaxAckPending {
-		t.Fatalf("queue depth %d fell back to the message brake %d", flowQueueDepth, flowMaxAckPending)
-	}
-	if got := cap(testFlowSubscriber().queue); got != flowQueueDepth {
-		t.Fatalf("bound lane queue = %d, want the derived depth %d", got, flowQueueDepth)
-	}
-}
-
-// The server heartbeats a STALLED consumer forever, so heartbeat silence can
-// never expose an unanswered flow-control conversation: lastControl stays fresh
-// while delivery is frozen. The streak is the only signal that survives that.
-func TestStalledHeartbeatStreakIsReportedAsAWedge(t *testing.T) {
-	sub := testFlowSubscriber()
-
-	for i := 0; i < flowWedgeStreak-1; i++ {
-		sub.handleStatus(stalledFlowStatus(), controlStatus)
-	}
-	if sub.wedged.Load() != 0 {
-		t.Fatalf("wedge reported after %d stalls, before the streak threshold", flowWedgeStreak-1)
-	}
-
-	sub.handleStatus(stalledFlowStatus(), controlStatus)
-	if sub.wedged.Load() != 1 {
-		t.Fatalf("wedged counter = %d after %d consecutive stalls, want 1",
-			sub.wedged.Load(), flowWedgeStreak)
-	}
-	// Reported, never repaired: recreating the consumer cannot grant a denied
-	// $JS.FC publish permission, and would churn the lane once a second.
-	if sub.recovering.Load() {
-		t.Fatal("a wedge triggered a consumer re-provision")
-	}
-}
-
-func TestADeliveryClearsTheWedgeStreak(t *testing.T) {
-	sub := testFlowSubscriber()
-	for i := 0; i < flowWedgeStreak-1; i++ {
-		sub.handleStatus(stalledFlowStatus(), controlStatus)
-	}
-
-	sub.deliveryCallback(laneDelivery("logical-id", []byte(`{"text":"hello"}`)))
-	if got := sub.stallStreak.Load(); got != 0 {
-		t.Fatalf("streak = %d after a delivery, want it cleared", got)
-	}
-
-	sub.handleStatus(stalledFlowStatus(), controlStatus)
-	if sub.wedged.Load() != 0 {
-		t.Fatalf("a stall after a delivery completed the old streak: wedged=%d", sub.wedged.Load())
-	}
-}
-
-// An unstalled heartbeat is the healthy case and must not accumulate anything:
-// the server sends one every second for the life of a quiet lane.
-func TestQuietHeartbeatsNeverAccumulateAWedge(t *testing.T) {
-	sub := testFlowSubscriber()
-	for i := 0; i < 4*flowWedgeStreak; i++ {
-		sub.handleStatus(flowStatus("Idle Heartbeat"), controlStatus)
-	}
-	if sub.stallStreak.Load() != 0 || sub.wedged.Load() != 0 {
-		t.Fatalf("quiet heartbeats counted as stalls: streak=%d wedged=%d",
-			sub.stallStreak.Load(), sub.wedged.Load())
-	}
-}
-
-func TestHeartbeatSilenceIsWhatDetectsADeletedConsumer(t *testing.T) {
-	now := time.Now()
-	if heartbeatLost(now, now.Add(-2*time.Second).UnixNano()) {
-		t.Fatal("a live one-second heartbeat was treated as silence")
-	}
-	if !heartbeatLost(now, now.Add(-4*time.Second).UnixNano()) {
-		t.Fatal("three missed heartbeats did not trigger recovery")
-	}
-}
-
-func TestStatusesOtherThanControlNeverReprovision(t *testing.T) {
-	sub := testFlowSubscriber()
-	// 2.14.3 publishes nothing but 100-class control messages to a push delivery
-	// subject, so a 409 here is noise and must not destroy the binding.
-	deleted := flowStatus("Consumer Deleted")
-	deleted.Header.Set(statusHeader, "409")
-	sub.handleStatus(deleted, "409")
-
-	if sub.recovering.Load() {
-		t.Fatal("a 409 status triggered a re-provision")
-	}
-}
-
-func TestDeliveryGapIsBoundedByTheAckWindow(t *testing.T) {
-	// Everything inside MaxAckPending is legitimately in flight.
-	if deliveriesWereLost(flowMaxAckPending, 0) {
-		t.Fatal("a full in-flight window was reported as loss")
-	}
-	if !deliveriesWereLost(flowMaxAckPending+1, 0) {
-		t.Fatal("deliveries past the ack window were not reported as lost")
-	}
-}
-
-func TestFlowSubscriberRejectsForeignSubjects(t *testing.T) {
-	sub := testFlowSubscriber()
-	if _, err := sub.Subscribe(context.Background(), "twitch.ingress.event.premium"); err == nil {
-		t.Fatal("flow subscriber accepted a subject it is not bound to")
-	}
-}
-
-func TestEveryUnitSharesOnePodLaneChannel(t *testing.T) {
-	sub := testFlowSubscriber()
-	first, err := sub.Subscribe(context.Background(), sub.subject)
-	if err != nil {
-		t.Fatal(err)
-	}
-	second, err := sub.Subscribe(context.Background(), sub.subject)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// One consumer per pod means one delivery stream per pod: the units compete
-	// for the same channel instead of each binding its own consumer.
-	if first != second {
-		t.Fatal("a second consumer unit was handed its own lane channel")
 	}
 }
 
@@ -706,61 +233,25 @@ func TestPartitionedLanesBindReceiptLevelConsumers(t *testing.T) {
 				"twitch.ingress.event.standard": TwitchIngressStandardStream.Name,
 				"twitch.ingress.event.premium":  TwitchIngressStream.Name,
 			} {
-				target, err := targetForTopic(subject)
-				if err != nil {
-					t.Fatalf("targetForTopic(%q): %v", subject, err)
-				}
-				if target.stream != stream {
-					t.Fatalf("%q binds stream %q, want %q", subject, target.stream, stream)
-				}
-				if got := subscriber.laneModeFor(target); got != mode.want {
-					t.Fatalf("%q on %q got mode %q, want %q", subject, target.stream, got, mode.want)
-				}
+				requireLaneBinding(t, subscriber, subject, stream, mode.want)
 			}
 		})
 	}
 }
 
-func flowStatus(description string) *nats.Msg {
-	control := nats.NewMsg("_INBOX.BAGEL.worker_premium")
-	control.Header.Set(statusHeader, controlStatus)
-	control.Header.Set(descriptionHeader, description)
-	return control
-}
-
-// stalledFlowStatus is the heartbeat a server sends while it is holding the
-// delivery window shut: the stalled marker carries the flow-control id the
-// response has to go to.
-func stalledFlowStatus() *nats.Msg {
-	stalled := flowStatus("Idle Heartbeat")
-	stalled.Header.Set(consumerStalledHeader, "$JS.FC.TWITCH_INGRESS.worker_standard.stall")
-	return stalled
-}
-
-func laneDelivery(id string, payload []byte) *nats.Msg {
-	wire := nats.NewMsg("twitch.ingress.event.standard")
-	wire.Header.Set(MessageIDHeader, id)
-	wire.Data = payload
-	return wire
-}
-
-func mustFlowMessage(t *testing.T, wire *nats.Msg) *Message {
+// requireLaneBinding walks one subject through the binding a consumer
+// provisions against: the stream it resolves to and the acknowledgement
+// contract its lane gets.
+func requireLaneBinding(t *testing.T, subscriber *fleetSubscriber, subject, stream string, want laneConsumeMode) {
 	t.Helper()
-	msg, err := messageFromNATS(wire)
+	target, err := targetForTopic(subject)
 	if err != nil {
-		t.Fatal(err)
+		t.Fatalf("targetForTopic(%q): %v", subject, err)
 	}
-	return msg
-}
-
-func testFlowSubscriber() *flowSubscriber {
-	return &flowSubscriber{
-		stream: TwitchIngressStream.Name, subject: "twitch.ingress.event.standard",
-		group: "worker", consumer: "worker_twitch_ingress_event_standard_pod_1",
-		log:   zap.NewNop(),
-		queue: make(chan flowDelivery, flowQueueDepth),
-		// Unbuffered, like the real lane: nothing reads it in these tests.
-		output:  make(chan *Message),
-		closeCh: make(chan struct{}),
+	if target.stream != stream {
+		t.Fatalf("%q binds stream %q, want %q", subject, target.stream, stream)
+	}
+	if got := subscriber.laneModeFor(target); got != want {
+		t.Fatalf("%q on %q got mode %q, want %q", subject, target.stream, got, want)
 	}
 }
