+++ pkg/bus/orbit_batch_test.go
@@ -1,703 +1,14 @@
 package bus
 
 import (
-	"context"
-	"errors"
-	"fmt"
 	"strconv"
-	"strings"
 	"testing"
 	"time"
 
 	"github.com/nats-io/nats.go"
-	jsapi "github.com/nats-io/nats.go/jetstream"
 	"github.com/synadia-io/orbit.go/jetstreamext"
 	"go.uber.org/zap"
 )
-
-// NATS_PUBLISH_WIRE is set in no manifest under deploy/, so this default is the
-// whole fleet's behaviour: every Go publisher takes it the moment its image
-// rolls. It must therefore be the wire that is already running. A batching
-// default would flip sesame, outgress, users, projector and the data services in
-// one deploy, and would spend the same per-stream budget of 50 open batches that
-// the Elixir ingress fleet sizes itself against.
-func TestPublishWireModeDefaultsToSingle(t *testing.T) {
-	t.Setenv("NATS_PUBLISH_WIRE", "")
-	if publishWireMode() != wireSingle {
-		t.Fatal("unset NATS_PUBLISH_WIRE must select the per-message PubAck wire")
-	}
-}
-
-// A typo must degrade, not escalate. Both batching wires have a wider blast
-// radius than the default — an ambiguous atomic outcome costs a whole cohort
-// where the single wire costs one message — so an unparseable value resolves to
-// the smallest one rather than to the configured default.
-func TestPublishWireModeFallsBackToTheSmallestBlastRadius(t *testing.T) {
-	for _, value := range []string{"nonsense", "Atomic", "ATOMIC", "atomic ", "1"} {
-		t.Setenv("NATS_PUBLISH_WIRE", value)
-		if got := publishWireMode(); got != wireSingle {
-			t.Errorf("NATS_PUBLISH_WIRE=%q selected %v; an unrecognised value must fall back to the per-message wire", value, got)
-		}
-	}
-}
-
-func TestPublishWireModeSelectsExplicitWires(t *testing.T) {
-	t.Setenv("NATS_PUBLISH_WIRE", "atomic")
-	if publishWireMode() != wireAtomic {
-		t.Fatal("NATS_PUBLISH_WIRE=atomic must select the ADR-050 batch wire")
-	}
-	t.Setenv("NATS_PUBLISH_WIRE", "single")
-	if publishWireMode() != wireSingle {
-		t.Fatal("NATS_PUBLISH_WIRE=single must select the per-message PubAck wire")
-	}
-	t.Setenv("NATS_PUBLISH_WIRE", "fast")
-	if publishWireMode() != wireFast {
-		t.Fatal("NATS_PUBLISH_WIRE=fast must select the Fast-Ingest wire")
-	}
-}
-
-func TestCohortWireBypassesBatchingForLoneMessage(t *testing.T) {
-	if got := cohortWire(wireFast, 1); got != wireSingle {
-		t.Fatalf("one-message fast cohort = %v, want the plain async wire", got)
-	}
-	if got := cohortWire(wireAtomic, 1); got != wireSingle {
-		t.Fatalf("one-message atomic cohort = %v, want the plain async wire", got)
-	}
-	if got := cohortWire(wireFast, 2); got != wireFast {
-		t.Fatalf("fast cohort = %v, want the Fast-Ingest wire", got)
-	}
-	if got := cohortWire(wireAtomic, 2); got != wireAtomic {
-		t.Fatalf("atomic cohort = %v, want the ADR-050 wire", got)
-	}
-	if got := cohortWire(wireSingle, 128); got != wireSingle {
-		t.Fatalf("single-wire cohort = %v, want the plain async wire", got)
-	}
-}
-
-func TestFastPublishSettingsAreBounded(t *testing.T) {
-	t.Setenv("NATS_FAST_PUBLISH_BATCH_SIZE", "0")
-	t.Setenv("NATS_FAST_PUBLISH_FLOW", "999999")
-	t.Setenv("NATS_FAST_PUBLISH_OUTSTANDING_ACKS", "-1")
-
-	if got := publishBatchSize(wireFast); got != 1 {
-		t.Fatalf("fast batch size = %d, want lower bound 1", got)
-	}
-	if got := fastPublishFlow(1000, 8); got != 124 {
-		t.Fatalf("fast flow = %d, want useful cohort ceiling 124", got)
-	}
-	if got := fastPublishOutstanding(); got != 1 {
-		t.Fatalf("fast outstanding = %d, want lower bound 1", got)
-	}
-	if got := publishBatchSize(wireSingle); got != defaultPublishBatchSize {
-		t.Fatalf("single batch size = %d, want %d", got, defaultPublishBatchSize)
-	}
-}
-
-func TestFastPublishDefaultsToLongSession(t *testing.T) {
-	t.Setenv("NATS_FAST_PUBLISH_BATCH_SIZE", "")
-	if got := publishBatchSize(wireFast); got != 8192 {
-		t.Fatalf("default fast session size = %d, want 8192", got)
-	}
-}
-
-// The broker enforces no maximum fast-batch size, so the clamp is a chosen
-// blast-radius bound rather than a protocol ceiling; it still has to bind.
-func TestFastPublishSessionSizeIsClampedToTheChosenBound(t *testing.T) {
-	t.Setenv("NATS_FAST_PUBLISH_BATCH_SIZE", "200000")
-	if got := publishBatchSize(wireFast); got != fastSessionMax {
-		t.Fatalf("fast session size = %d, want the chosen bound %d", got, fastSessionMax)
-	}
-}
-
-func TestFastPublishBatchWaitIsBounded(t *testing.T) {
-	t.Setenv("NATS_FAST_PUBLISH_BATCH_WAIT", "500ms")
-	if got := publishBatchWait(wireFast); got != 100*time.Millisecond {
-		t.Fatalf("fast batch wait = %s, want upper bound 100ms", got)
-	}
-	if got := publishBatchWait(wireSingle); got != defaultPublishBatchWait {
-		t.Fatalf("single batch wait = %s, want %s", got, defaultPublishBatchWait)
-	}
-}
-
-func TestFastCohortCommitsWholeSession(t *testing.T) {
-	publisher := &stubFastPublisher{ack: &jetstreamext.BatchAck{BatchSize: 3, Sequence: 91}}
-
-	outcome := publishFastCohort(publisher, stagedBatch(3), &firstAsyncPublishError{})
-
-	if outcome.err != nil {
-		t.Fatalf("complete fast cohort failed: %v", outcome.err)
-	}
-	if outcome.acked != 3 {
-		t.Fatalf("committed session acknowledged %d/3 messages", outcome.acked)
-	}
-	if publisher.addCalls != 2 || publisher.commitCalls != 1 {
-		t.Fatalf("fast session sent %d adds and %d commits, want 2 and 1", publisher.addCalls, publisher.commitCalls)
-	}
-	if publisher.closeCalls != 0 {
-		t.Fatalf("a committed session was closed %d times", publisher.closeCalls)
-	}
-}
-
-func TestFastCohortDoesNotCommitAfterAddFailure(t *testing.T) {
-	want := errors.New("add failed")
-	publisher := &stubFastPublisher{addErr: want}
-
-	outcome := publishFastCohort(publisher, stagedBatch(2), &firstAsyncPublishError{})
-
-	if !errors.Is(outcome.err, want) {
-		t.Fatalf("publishFastCohort() error = %v, want %v", outcome.err, want)
-	}
-	if publisher.commitCalls != 0 {
-		t.Fatalf("commit called %d times after Add failure", publisher.commitCalls)
-	}
-}
-
-func TestFastCohortStopsAtFirstAsyncError(t *testing.T) {
-	asyncErr := &firstAsyncPublishError{}
-	want := errors.New("gap detected")
-	publisher := &stubFastPublisher{ack: &jetstreamext.BatchAck{BatchSize: 3}, onAdd: func() { asyncErr.set(want) }}
-
-	outcome := publishFastCohort(publisher, stagedBatch(3), asyncErr)
-
-	if !errors.Is(outcome.err, want) {
-		t.Fatalf("publishFastCohort() error = %v, want the reported session error", outcome.err)
-	}
-	if publisher.commitCalls != 0 {
-		t.Fatalf("commit called %d times after an out-of-band session error", publisher.commitCalls)
-	}
-}
-
-func TestFastCohortFailsOnAsyncErrorReportedAtCommit(t *testing.T) {
-	asyncErr := &firstAsyncPublishError{}
-	want := errors.New("sequence error")
-	publisher := &stubFastPublisher{ack: &jetstreamext.BatchAck{BatchSize: 2}, onCommit: func() { asyncErr.set(want) }}
-
-	if outcome := publishFastCohort(publisher, stagedBatch(2), asyncErr); !errors.Is(outcome.err, want) {
-		t.Fatalf("publishFastCohort() error = %v, want the reported session error", outcome.err)
-	}
-}
-
-// The fast terminal count is the highest batch sequence the broker accepted, so
-// a short one proves only that the session did not cover every sequence sent.
-// What landed comes from the flow acknowledgements, and the report must say both.
-func TestFastCohortShortCommitAckReportsCoverageAndStoredPrefix(t *testing.T) {
-	publisher := &stubFastPublisher{
-		ack:    &jetstreamext.BatchAck{BatchSize: 2, Sequence: 404},
-		addAck: fixedAckSequences(0, 1),
-	}
-
-	outcome := publishFastCohort(publisher, stagedBatch(3), &firstAsyncPublishError{})
-
-	if outcome.err == nil {
-		t.Fatal("a commit ack covering fewer sequences must fail the cohort")
-	}
-	if outcome.acked != 1 {
-		t.Fatalf("stored prefix = %d, want the acknowledged flow sequence 1", outcome.acked)
-	}
-	for _, want := range []string{"2/3 batch sequences", "1 messages acknowledged as stored", "stream sequence 404"} {
-		if !strings.Contains(outcome.err.Error(), want) {
-			t.Fatalf("short-ack error %q does not report %q", outcome.err, want)
-		}
-	}
-}
-
-func TestFastCohortFailsOnMissingCommitAck(t *testing.T) {
-	publisher := &stubFastPublisher{}
-	if outcome := publishFastCohort(publisher, stagedBatch(2), &firstAsyncPublishError{}); outcome.err == nil {
-		t.Fatal("a missing commit ack must fail the cohort")
-	}
-}
-
-// Fast-Ingest persists on arrival, so a dead session leaves a stored prefix.
-// The prefix the broker acknowledged must survive the failure verdict.
-func TestFastCohortKeepsTheAcknowledgedPrefix(t *testing.T) {
-	publisher := &stubFastPublisher{
-		commitErr: errors.New("ack timeout"),
-		addAck:    fixedAckSequences(0, 0, 0, 3, 3, 3, 6),
-	}
-
-	outcome := publishFastCohort(publisher, stagedBatch(8), &firstAsyncPublishError{})
-
-	if outcome.err == nil {
-		t.Fatal("a failed commit must fail the unacknowledged suffix")
-	}
-	if outcome.acked != 6 {
-		t.Fatalf("stored prefix = %d, want the highest acknowledged sequence 6", outcome.acked)
-	}
-	if outcome.rejectedFirst {
-		t.Fatal("a session that stored a prefix must never be replayed")
-	}
-}
-
-// Orbit never writes the initialErrCh it reads, so a rejection of the session's
-// first message reaches only the error handler while AddMsg times out. The typed
-// rejection is what has to be reported and what makes the replay safe.
-func TestFastCohortPrefersTheTypedRejectionOverTheAckTimeout(t *testing.T) {
-	rejected := &jsapi.APIError{Code: 400, ErrorCode: 10205, Description: "batch publish is disabled"}
-	asyncErr := &firstAsyncPublishError{}
-	asyncErr.set(rejected)
-	publisher := &stubFastPublisher{addErr: errors.New("batch message 1 ack timeout")}
-
-	outcome := publishFastCohort(publisher, stagedBatch(4), asyncErr)
-
-	if !errors.Is(outcome.err, error(rejected)) {
-		t.Fatalf("publishFastCohort() error = %v, want the typed broker rejection", outcome.err)
-	}
-	if !outcome.rejectedFirst {
-		t.Fatal("a definite rejection of message 1 must permit the per-message wire")
-	}
-}
-
-// Past the first message a typed error no longer proves nothing was stored:
-// everything the broker already took is in the stream and visible to consumers.
-func TestFastCohortDoesNotReplayPastTheFirstMessage(t *testing.T) {
-	rejected := &jsapi.APIError{Code: 400, ErrorCode: 10208, Description: "batch publish ID unknown"}
-	asyncErr := &firstAsyncPublishError{}
-	publisher := &stubFastPublisher{addAck: fixedAckSequences(0, 1)}
-	publisher.onAdd = func() {
-		if publisher.addCalls == 2 {
-			asyncErr.set(rejected)
-		}
-	}
-
-	outcome := publishFastCohort(publisher, stagedBatch(4), asyncErr)
-
-	if !errors.Is(outcome.err, error(rejected)) {
-		t.Fatalf("publishFastCohort() error = %v, want the reported rejection", outcome.err)
-	}
-	if outcome.rejectedFirst {
-		t.Fatal("a rejection after the first message must not permit a replay")
-	}
-}
-
-// Orbit's FastPublisher has no Discard: without a Close the ack-inbox
-// subscription and the broker's per-stream inflight slot survive every abandoned
-// cohort, so each path that leaves the session unfinished has to close it once.
-func TestFastCohortClosesTheSessionOnEveryAbortPath(t *testing.T) {
-	asyncErr := &firstAsyncPublishError{}
-	reported := &stubFastPublisher{ack: &jetstreamext.BatchAck{BatchSize: 3}}
-	reported.onAdd = func() { asyncErr.set(errors.New("gap detected")) }
-
-	for name, session := range map[string]struct {
-		publisher *stubFastPublisher
-		asyncErr  *firstAsyncPublishError
-	}{
-		"add error":         {publisher: &stubFastPublisher{addErr: errors.New("publish failed")}},
-		"commit error":      {publisher: &stubFastPublisher{commitErr: errors.New("ack timeout")}},
-		"out-of-band error": {publisher: reported, asyncErr: asyncErr},
-	} {
-		errs := session.asyncErr
-		if errs == nil {
-			errs = &firstAsyncPublishError{}
-		}
-		outcome := publishFastCohort(session.publisher, stagedBatch(3), errs)
-		if outcome.err == nil {
-			t.Fatalf("%s: expected a failed cohort", name)
-		}
-		if session.publisher.closeCalls != 1 {
-			t.Fatalf("%s: session closed %d times, want exactly 1", name, session.publisher.closeCalls)
-		}
-	}
-}
-
-// A session that reached its commit is already finished on Orbit's side: the
-// commit unsubscribes the ack inbox whatever the broker answered, so a short
-// terminal ack is a reporting problem, not a leak, and must not send a second
-// end-of-batch.
-func TestFastCohortDoesNotCloseACommittedSession(t *testing.T) {
-	publisher := &stubFastPublisher{ack: &jetstreamext.BatchAck{BatchSize: 2}}
-
-	outcome := publishFastCohort(publisher, stagedBatch(3), &firstAsyncPublishError{})
-
-	if outcome.err == nil {
-		t.Fatal("a commit ack covering fewer sequences must fail the cohort")
-	}
-	if publisher.closeCalls != 0 {
-		t.Fatalf("committed session closed %d times", publisher.closeCalls)
-	}
-}
-
-func TestFastCohortStoredPrefixNeverClaimsAFailedCohort(t *testing.T) {
-	committed := fastCohortOutcome{acked: 4}
-	if got := committed.storedPrefix(4); got != 4 {
-		t.Fatalf("committed prefix = %d, want the whole cohort", got)
-	}
-	failed := fastCohortOutcome{acked: 4, err: errors.New("ack timeout")}
-	if got := failed.storedPrefix(4); got != 3 {
-		t.Fatalf("failed prefix = %d, want at most one short of the cohort", got)
-	}
-	unknown := fastCohortOutcome{err: errors.New("ack timeout")}
-	if got := unknown.storedPrefix(4); got != 0 {
-		t.Fatalf("unacknowledged prefix = %d, want none", got)
-	}
-}
-
-// A failed Fast-Ingest cohort has two verdicts, not one: the acknowledged prefix
-// is durable and a confirmed caller must not retry it.
-func TestFinishFastSplitsTheAcknowledgedPrefixFromTheFailedSuffix(t *testing.T) {
-	worker := &publishBatchWorker{owner: newTestBatchPublisher()}
-	batch := confirmedBatch(5)
-	want := errors.New("session aborted")
-
-	worker.finishFast(batch, fastCohortOutcome{acked: 2, err: want})
-
-	for i, request := range batch {
-		got := <-request.confirmed
-		if i < 2 && got != nil {
-			t.Fatalf("message %d of the acknowledged prefix failed: %v", i, got)
-		}
-		if i >= 2 && !errors.Is(got, want) {
-			t.Fatalf("message %d of the unacknowledged suffix = %v, want %v", i, got, want)
-		}
-	}
-	if completed := worker.owner.completed; completed != uint64(len(batch)) {
-		t.Fatalf("resolved %d/%d messages", completed, len(batch))
-	}
-}
-
-func TestFastFallbackReplaysOnlyADefiniteFirstMessageRejection(t *testing.T) {
-	rejected := &jsapi.APIError{Code: 400, ErrorCode: 10205, Description: "batch publish is disabled"}
-	if !brokerRejectedBatch(rejected) {
-		t.Fatal("a typed rejection of message 1 must permit the per-message wire")
-	}
-	outcome := fastCohortOutcome{}.abort(&stubFastPublisher{}, 0, rejected)
-	if !outcome.rejectedFirst {
-		t.Fatal("abort at position 0 with nothing acknowledged must permit a replay")
-	}
-	stored := fastCohortOutcome{acked: 1}.abort(&stubFastPublisher{}, 0, rejected)
-	if stored.rejectedFirst {
-		t.Fatal("an acknowledged prefix must block the replay even at position 0")
-	}
-}
-
-func TestAtomicFallbackRequiresExplicitBrokerRejection(t *testing.T) {
-	rejected := &jsapi.APIError{Code: 400, ErrorCode: 10174, Description: "batch publish not enabled"}
-	if !brokerRejectedBatch(rejected) {
-		t.Fatal("a typed API rejection must permit the per-message fallback")
-	}
-	if !brokerRejectedBatch(fmt.Errorf("orbit: %w", rejected)) {
-		t.Fatal("a wrapped API rejection must permit the per-message fallback")
-	}
-	if !brokerRejectedBatch(jetstreamext.ErrBatchPublishNotEnabled) {
-		t.Fatal("a server feature rejection must permit the per-message fallback")
-	}
-	for name, err := range map[string]error{
-		"transport":       errors.New("nats: connection closed"),
-		"timeout":         context.DeadlineExceeded,
-		"short ack":       errors.New("bus: Orbit atomic commit stored 2/3 messages"),
-		"undecodable ack": jetstreamext.ErrInvalidBatchAck,
-		"closed batch":    jetstreamext.ErrBatchClosed,
-	} {
-		if brokerRejectedBatch(err) {
-			t.Fatalf("%s ambiguity must not permit replay: %v", name, err)
-		}
-	}
-}
-
-func TestAtomicFallbackDoesNotReplayAmbiguousFailure(t *testing.T) {
-	// A worker with no connection state proves the ambiguous path never reaches
-	// the re-publish machinery.
-	worker := &publishBatchWorker{}
-	want := errors.New("nats: connection closed")
-	if err := worker.atomicFallback(stagedBatch(3), want); !errors.Is(err, want) {
-		t.Fatalf("atomicFallback() error = %v, want the original cause", err)
-	}
-}
-
-func TestAtomicPublisherRejectsOversizeCohortClientSide(t *testing.T) {
-	worker := &publishBatchWorker{}
-	if err := worker.publishAtomic(stagedBatch(atomicBatchMax + 1)); err == nil {
-		t.Fatal("oversize atomic cohort reached Orbit")
-	}
-	if err := worker.publishAtomic(nil); err == nil {
-		t.Fatal("empty atomic cohort reached Orbit")
-	}
-}
-
-// The cohort size is the only lever on the server's per-cohort RAFT proposal, so
-// it is tunable; the clamp is the ADR-050 protocol range, not a preference.
-func TestAtomicPublishCohortSizeIsBounded(t *testing.T) {
-	t.Setenv("NATS_ATOMIC_PUBLISH_BATCH_SIZE", "")
-	if got := publishBatchSize(wireAtomic); got != defaultAtomicPublishBatchSize {
-		t.Fatalf("default atomic cohort = %d, want %d", got, defaultAtomicPublishBatchSize)
-	}
-	for value, want := range map[string]int{
-		"1":       2,
-		"0":       2,
-		"-64":     2,
-		"5000":    atomicBatchMax,
-		"512":     512,
-		"garbage": defaultAtomicPublishBatchSize,
-	} {
-		t.Setenv("NATS_ATOMIC_PUBLISH_BATCH_SIZE", value)
-		if got := publishBatchSize(wireAtomic); got != want {
-			t.Fatalf("NATS_ATOMIC_PUBLISH_BATCH_SIZE=%s gives cohort %d, want %d", value, got, want)
-		}
-	}
-}
-
-func TestAtomicPublishBatchWaitIsBounded(t *testing.T) {
-	t.Setenv("NATS_ATOMIC_PUBLISH_BATCH_WAIT", "")
-	if got := publishBatchWait(wireAtomic); got != defaultPublishBatchWait {
-		t.Fatalf("default atomic batch wait = %s, want %s", got, defaultPublishBatchWait)
-	}
-	for value, want := range map[string]time.Duration{
-		"100us": 500 * time.Microsecond,
-		"0s":    500 * time.Microsecond,
-		"1s":    20 * time.Millisecond,
-		"8ms":   8 * time.Millisecond,
-	} {
-		t.Setenv("NATS_ATOMIC_PUBLISH_BATCH_WAIT", value)
-		if got := publishBatchWait(wireAtomic); got != want {
-			t.Fatalf("NATS_ATOMIC_PUBLISH_BATCH_WAIT=%s gives wait %s, want %s", value, got, want)
-		}
-	}
-}
-
-func TestAtomicPublishOverlapIsOnUnlessDisabled(t *testing.T) {
-	t.Setenv("NATS_ATOMIC_PUBLISH_OVERLAP", "")
-	if !atomicPublishOverlap() {
-		t.Fatal("the commit overlap must be on by default")
-	}
-	t.Setenv("NATS_ATOMIC_PUBLISH_OVERLAP", "false")
-	if atomicPublishOverlap() {
-		t.Fatal("NATS_ATOMIC_PUBLISH_OVERLAP=false must keep the commit on the worker")
-	}
-}
-
-// Orbit's commit is a publish: the cohort's last message rides CommitMsg. The
-// worker therefore stages exactly the first N-1 messages, in order, and the
-// commit carries the last one.
-func TestAtomicCohortStagesEveryMessageButTheCommitPayload(t *testing.T) {
-	publisher := &stubAtomicPublisher{}
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) { return publisher, nil })
-	batch := confirmedBatch(4)
-
-	worker.publishAtomicOverlapped(batch)
-	worker.acks.Wait()
-
-	for i := range batch {
-		if err := <-batch[i].confirmed; err != nil {
-			t.Fatalf("message %d of a committed cohort failed: %v", i, err)
-		}
-	}
-	if len(publisher.added) != 3 {
-		t.Fatalf("staged %d messages, want the cohort minus its commit payload", len(publisher.added))
-	}
-	for i, msg := range publisher.added {
-		if got := msg.Header.Get(messageIDHeader); got != "id-"+itoa(i+1) {
-			t.Fatalf("staged message %d is %s, want the cohort in wire order", i, got)
-		}
-	}
-	if publisher.committed != batch[3].msg {
-		t.Fatal("the commit did not carry the cohort's last message")
-	}
-}
-
-// The overlap's whole point is that a cohort no longer waits behind an older
-// cohort's commit ack. Each cohort resolves on its own commit, and its confirmed
-// callers hear about that commit and no other.
-func TestAtomicOverlapResolvesEachCohortOnItsOwnCommit(t *testing.T) {
-	slow, opened := make(chan struct{}), 0
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) {
-		opened++
-		if opened == 1 {
-			return &stubAtomicPublisher{release: slow}, nil
-		}
-		return &stubAtomicPublisher{}, nil
-	})
-	first, second := confirmedBatch(3), confirmedBatch(2)
-
-	worker.publishAtomicOverlapped(first)
-	worker.publishAtomicOverlapped(second)
-
-	for i := range second {
-		if err := <-second[i].confirmed; err != nil {
-			t.Fatalf("message %d of the second cohort failed: %v", i, err)
-		}
-	}
-	select {
-	case err := <-first[0].confirmed:
-		t.Fatalf("the first cohort resolved before its own commit: %v", err)
-	default:
-	}
-
-	close(slow)
-	worker.acks.Wait()
-	for i := range first {
-		if err := <-first[i].confirmed; err != nil {
-			t.Fatalf("message %d of the first cohort failed: %v", i, err)
-		}
-	}
-	if completed := worker.owner.completed; completed != 5 {
-		t.Fatalf("resolved %d/5 messages", completed)
-	}
-}
-
-// The overlap moves the ack wait, never the at-most-once rule: an ambiguous
-// commit failure fails the cohort where it stands. The worker has no JetStream
-// context that can publish, so a replay would be visible as a refused send.
-func TestAtomicOverlapDoesNotReplayAnAmbiguousCommitFailure(t *testing.T) {
-	want := errors.New("nats: connection closed")
-	replay := &stubJetStream{err: errors.New("the per-message wire must not be used")}
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) {
-		return &stubAtomicPublisher{commitErr: want}, nil
-	})
-	worker.js = replay
-	batch := confirmedBatch(3)
-
-	worker.publishAtomicOverlapped(batch)
-	worker.acks.Wait()
-
-	for i := range batch {
-		if err := <-batch[i].confirmed; !errors.Is(err, want) {
-			t.Fatalf("message %d = %v, want the ambiguous cause %v", i, err, want)
-		}
-	}
-	if replay.published != 0 {
-		t.Fatalf("an ambiguous commit failure re-published %d messages", replay.published)
-	}
-}
-
-// A typed broker rejection is the one commit failure that proves nothing was
-// stored, and the replay it authorises belongs to the goroutine that resolved
-// the cohort, not to the worker.
-func TestAtomicOverlapReplaysADefiniteBrokerRejection(t *testing.T) {
-	rejected := &jsapi.APIError{Code: 400, ErrorCode: 10174, Description: "batch publish not enabled"}
-	replay := &stubJetStream{err: errors.New("per-message wire refused")}
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) {
-		return &stubAtomicPublisher{commitErr: rejected}, nil
-	})
-	worker.js = replay
-	batch := confirmedBatch(3)
-
-	worker.publishAtomicOverlapped(batch)
-	worker.acks.Wait()
-
-	if replay.published == 0 {
-		t.Fatal("a typed broker rejection must reach the per-message wire")
-	}
-	for i := range batch {
-		if err := <-batch[i].confirmed; err == nil {
-			t.Fatalf("message %d passed although its replay was refused", i)
-		}
-	}
-}
-
-// A cohort Orbit refuses to stage never reaches a commit, so its verdict is
-// settled by the same rules on the same goroutine.
-func TestAtomicOverlapResolvesAStagingFailureWithoutCommitting(t *testing.T) {
-	want := errors.New("nats: connection closed")
-	publisher := &stubAtomicPublisher{addErr: want}
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) { return publisher, nil })
-	worker.js = &stubJetStream{err: errors.New("the per-message wire must not be used")}
-	batch := confirmedBatch(3)
-
-	worker.publishAtomicOverlapped(batch)
-	worker.acks.Wait()
-
-	for i := range batch {
-		if err := <-batch[i].confirmed; !errors.Is(err, want) {
-			t.Fatalf("message %d = %v, want the staging cause %v", i, err, want)
-		}
-	}
-	if publisher.committed != nil {
-		t.Fatal("a cohort that failed to stage was committed")
-	}
-	if publisher.discards != 1 {
-		t.Fatalf("a half-staged batch was discarded %d times, want exactly 1", publisher.discards)
-	}
-}
-
-// The slots are the client's share of the broker's cap on concurrently staged
-// batches per stream, so a worker must stall on staging rather than open a fifth
-// batch.
-func TestAtomicOverlapSlotsBoundTheCommitDepth(t *testing.T) {
-	release, started := make(chan struct{}), make(chan struct{}, 8)
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) {
-		return &stubAtomicPublisher{commitStarted: started, release: release}, nil
-	})
-	worker.slots = make(chan struct{}, 2)
-
-	worker.publishAtomicOverlapped(confirmedBatch(2))
-	worker.publishAtomicOverlapped(confirmedBatch(2))
-	awaitSignal(t, started, "the first two commits never started")
-	awaitSignal(t, started, "the second commit never started")
-
-	staged := make(chan struct{})
-	go func() {
-		worker.publishAtomicOverlapped(confirmedBatch(2))
-		close(staged)
-	}()
-	select {
-	case <-staged:
-		t.Fatal("a third cohort staged while both commit slots were held")
-	case <-started:
-		t.Fatal("a third commit started while both commit slots were held")
-	case <-time.After(50 * time.Millisecond):
-	}
-
-	close(release)
-	awaitSignal(t, staged, "the blocked cohort never staged after a slot was freed")
-	worker.acks.Wait()
-}
-
-// Stopping the worker must not abandon a cohort whose commit is still out: its
-// callers are waiting on that verdict and the connection is about to drain.
-func TestAtomicOverlapDrainsInFlightCommitsOnStop(t *testing.T) {
-	release, started := make(chan struct{}), make(chan struct{}, 4)
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) {
-		return &stubAtomicPublisher{commitStarted: started, release: release}, nil
-	})
-	worker.batchSize = 2
-	go worker.run()
-
-	batch := confirmedBatch(2)
-	worker.requests <- batch[0]
-	worker.requests <- batch[1]
-	awaitSignal(t, started, "the cohort never reached its commit")
-
-	close(worker.stop)
-	select {
-	case <-worker.done:
-		t.Fatal("the worker finished while a commit was still in flight")
-	case <-time.After(50 * time.Millisecond):
-	}
-
-	close(release)
-	awaitSignal(t, worker.done, "the worker did not finish once its commit resolved")
-	for i := range batch {
-		if err := <-batch[i].confirmed; err != nil {
-			t.Fatalf("message %d of a drained cohort failed: %v", i, err)
-		}
-	}
-}
-
-// With the overlap off the commit stays on the calling goroutine, which is what
-// keeps whole cohorts in stream order: the broker sequences an atomic batch at
-// commit, not on arrival.
-func TestAtomicCohortWithoutOverlapCommitsOnTheWorker(t *testing.T) {
-	publisher := &stubAtomicPublisher{}
-	worker := newOverlapWorker(func() (atomicCohortPublisher, error) { return publisher, nil })
-	worker.overlapCommit = false
-	batch := confirmedBatch(3)
-
-	worker.publishAtomicCohort(batch)
-
-	if publisher.committed == nil {
-		t.Fatal("the cohort returned to the worker before its commit")
-	}
-	for i := range batch {
-		select {
-		case err := <-batch[i].confirmed:
-			if err != nil {
-				t.Fatalf("message %d failed: %v", i, err)
-			}
-		default:
-			t.Fatalf("message %d was not resolved by the time the worker returned", i)
-		}
-	}
-}
 
 func TestStripOrbitBatchFramingKeepsFleetIdentity(t *testing.T) {
 	batch := framedBatch(2)
@@ -739,155 +50,6 @@ func TestCollectBatchStopsAtTheWireWait(t *testing.T) {
 	batch, ok := worker.collectBatch(stagedBatch(1)[0])
 	if !ok || len(batch) != 1 {
 		t.Fatalf("collectBatch() = (%d, %v), want the single staged message", len(batch), ok)
-	}
-}
-
-// newOverlapWorker builds the worker state an atomic cohort touches, with the
-// Orbit session replaced by a stub. Cohorts are handed to it directly, so the
-// stub is only ever reached from the goroutine that owns the cohort.
-func newOverlapWorker(newAtomic func() (atomicCohortPublisher, error)) *publishBatchWorker {
-	owner := newTestBatchPublisher()
-	owner.wire = wireAtomic
-	return &publishBatchWorker{
-		owner:         owner,
-		requests:      make(chan publishRequest, 8),
-		stop:          make(chan struct{}),
-		done:          make(chan struct{}),
-		slots:         make(chan struct{}, maxInflightCohorts),
-		batchSize:     defaultAtomicPublishBatchSize,
-		batchWait:     defaultPublishBatchWait,
-		overlapCommit: true,
-		newAtomic:     newAtomic,
-	}
-}
-
-func awaitSignal[T any](t *testing.T, signal <-chan T, message string) {
-	t.Helper()
-	select {
-	case <-signal:
-	case <-time.After(2 * time.Second):
-		t.Fatal(message)
-	}
-}
-
-// stubAtomicPublisher stands in for Orbit's BatchPublisher. release holds the
-// commit open so a cohort can be observed while its ack is still outstanding,
-// which is the state the overlap exists to create.
-type stubAtomicPublisher struct {
-	added         []*nats.Msg
-	committed     *nats.Msg
-	addErr        error
-	commitErr     error
-	discards      int
-	commitStarted chan struct{}
-	release       chan struct{}
-}
-
-func (p *stubAtomicPublisher) AddMsg(msg *nats.Msg, _ ...jetstreamext.BatchMsgOpt) error {
-	if p.addErr != nil {
-		return p.addErr
-	}
-	p.added = append(p.added, msg)
-	return nil
-}
-
-func (p *stubAtomicPublisher) CommitMsg(ctx context.Context, msg *nats.Msg, _ ...jetstreamext.BatchMsgOpt) (*jetstreamext.BatchAck, error) {
-	p.committed = msg
-	if p.commitStarted != nil {
-		p.commitStarted <- struct{}{}
-	}
-	if err := p.awaitRelease(ctx); err != nil {
-		return nil, err
-	}
-	if p.commitErr != nil {
-		return nil, p.commitErr
-	}
-	// The ordinary broker answer: every staged message plus the commit payload
-	// was stored.
-	return &jetstreamext.BatchAck{BatchSize: uint64(len(p.added) + 1), Sequence: 42}, nil
-}
-
-func (p *stubAtomicPublisher) awaitRelease(ctx context.Context) error {
-	if p.release == nil {
-		return nil
-	}
-	select {
-	case <-p.release:
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-func (p *stubAtomicPublisher) Discard() error {
-	p.discards++
-	return nil
-}
-
-// stubJetStream stands in for the per-message wire so a fallback replay is
-// observable without a broker. Only PublishMsgAsync is reachable from the
-// fallback; the embedded interface is nil and every other method would panic,
-// which is the point. err must be set: a nil PubAckFuture is not awaitable.
-type stubJetStream struct {
-	nats.JetStreamContext
-	published int
-	err       error
-}
-
-func (s *stubJetStream) PublishMsgAsync(*nats.Msg, ...nats.PubOpt) (nats.PubAckFuture, error) {
-	s.published++
-	return nil, s.err
-}
-
-// stubFastPublisher stands in for Orbit's FastPublisher so session failure
-// handling is exercised without a broker.
-type stubFastPublisher struct {
-	addErr      error
-	commitErr   error
-	ack         *jetstreamext.BatchAck
-	addAck      func(call int) *jetstreamext.FastPubAck
-	addCalls    int
-	commitCalls int
-	closeCalls  int
-	onAdd       func()
-	onCommit    func()
-}
-
-func (p *stubFastPublisher) AddMsg(*nats.Msg, ...jetstreamext.BatchMsgOpt) (*jetstreamext.FastPubAck, error) {
-	p.addCalls++
-	if p.onAdd != nil {
-		p.onAdd()
-	}
-	if p.addAck != nil {
-		return p.addAck(p.addCalls), p.addErr
-	}
-	return &jetstreamext.FastPubAck{}, p.addErr
-}
-
-func (p *stubFastPublisher) CommitMsg(context.Context, *nats.Msg, ...jetstreamext.BatchMsgOpt) (*jetstreamext.BatchAck, error) {
-	p.commitCalls++
-	if p.onCommit != nil {
-		p.onCommit()
-	}
-	return p.ack, p.commitErr
-}
-
-// Close answers the way Orbit does once a session is over: nothing left to
-// commit. What the seam needs is that it was called at all.
-func (p *stubFastPublisher) Close(context.Context) (*jetstreamext.BatchAck, error) {
-	p.closeCalls++
-	return nil, jetstreamext.ErrBatchClosed
-}
-
-// fixedAckSequences replays the flow-ack sequence Orbit reports per Add. The
-// server acknowledges the first message with sequence 0 and then only every
-// flow-th message, so the series is flat and lags the messages sent.
-func fixedAckSequences(sequences ...uint64) func(call int) *jetstreamext.FastPubAck {
-	return func(call int) *jetstreamext.FastPubAck {
-		if call > len(sequences) {
-			return &jetstreamext.FastPubAck{BatchSequence: uint64(call)}
-		}
-		return &jetstreamext.FastPubAck{BatchSequence: uint64(call), AckSequence: sequences[call-1]}
 	}
 }
